### 2023-11-26 #6092 @dfm
@athob — Thanks for your submission! Given the small size of the main code base, I have submitted this pre-review for editorial review. This usually takes a week or two and I'll report back here as soon as I hear back from the rest of the editorial board.

Should we also be reviewing https://github.com/athob/py-EnBiD-ananke and https://github.com/athob/py-Galaxia-ananke? Please feel free to provide any extra context that would be useful for the team when looking at the [scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) of this submission. Thanks!
