### 2023-11-26 #6092 @athob
Hi @dfm, thanks for your message!

Indeed the submodule repositories of [`py-EnBiD-ananke`](https://github.com/athob/py-EnBiD-ananke) and [`py-Galaxia-ananke`](https://github.com/athob/py-Galaxia-ananke) should be considered part of this submission as well.

Additionally, I would like to bring to your attention to some extent the submodule repository [`galaxia-ananke`](https://github.com/athob/galaxia-ananke) in [`py-Galaxia-ananke`](https://github.com/athob/py-Galaxia-ananke). The reason I say "to some extent" is that the state of the codebase at commit [295326d](https://github.com/athob/galaxia-ananke/commit/295326dd44c15679af5de1de1d32dbe1976a0808) represents the software used by co-author Robyn Sanderson in the original `ananke` framework from [Sanderson et al. 2020](https://ui.adsabs.harvard.edu/abs/2020ApJS..246....6Sl) (with the addition of LSST isochrones by co-author Nicolás Garavito-Camargo, addition that should be considered part of this submission). That software itself was also a modified iteration of the original `Galaxia` software by co-author Sanjib Sharma ([Sharma et al. 2011](https://ui.adsabs.harvard.edu/abs/2011ascl.soft01007S)). That said, all the [modifications made since commit 295326d](https://github.com/athob/galaxia-ananke/compare/295326d..e6ca844) are contributions that can be considered integral to this submission.

While reviewing the criteria outlined in the [Substantial Scholarly Effort Guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort), I noticed that the last criterion evaluates the likelihood of `py-ananke` getting cited. I have addressed this aspect in a dedicated section titled 'Past and Ongoing Applications' within the paper. Several ongoing projects are currently utilizing this new software and are committed to citing it. It's important to note that `py-ananke` was specifically designed to address usability inefficiencies present in the original `ananke` framework introduced by [Sanderson et al. in 2020](https://ui.adsabs.harvard.edu/abs/2020ApJS..246....6Sl). The usability challenges associated with the original framework have necessitated significant efforts in past projects utilizing `ananke`, which would have been alleviated had `py-ananke` been available then.

I would appreciate your feedback on the aforementioned points. If there are any additional inquiries or clarifications needed, please feel free to let me know. Thanks for your time and consideration, I am looking forward to hearing back from you after the editorial board's response.
