### 2023-12-13 #6092 @warrickball
Hi @athob,

Thanks for the explanation. We've agreed that between this principal repo and the two submodules, the code is substantial enough to proceed with the review. I'll start looking for reviewers. For simplicity, I'll say that the repos to be considered part of the review are:

* [py-ananke](https://github.com/athob/py-ananke)
* [py-EnBiD-ananke](https://github.com/athob/py-EnBiD-ananke)
* [py-Galaxia-ananke](https://github.com/athob/py-Galaxia-ananke)
