### 2024-01-17 #6234 @editorialbot
[REVIEW]: Generating synthetic star catalogs from simulated data for next-gen observatories with py-ananke
**Submitting author:** @athob (<a href="http://orcid.org/0000-0001-7928-1973">Adrien C. R. Thob</a>)
**Repository:** https://github.com/athob/py-ananke
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.0
