### 2024-01-17 #6234 @warrickball
Hi @lheckmann, @rrjbca, and thanks again for agreeing to review. This is the review thread for the paper. All of our correspondence will now happen here.

Please read the "Reviewer instructions & questions" in the [first comment](https://github.com/openjournals/joss-reviews/issues/6234#issue-2085768651) above, and generate your checklists by commenting `@editorialbot generate my checklist` on this issue. As you go over the submission, please check any items that you feel have been satisfied. There are also links to the JOSS reviewer guidelines.

The JOSS review is different from most other journals. We aim to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. We also encourage reviewers to submit issues and pull requests on the software repository. When doing so, please mention `openjournals/joss-reviews#6234` so that the issue/PR is linked to this thread. Please also feel free to comment and ask questions on this thread. JOSS editors have found it better to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire package.

We aim for the review process to be completed within about 4-6 weeks but start whenever you can. JOSS reviews are iterative and the authors can start responding while you continue to review other parts of the submission.

You're welcome to also assign yourself to this issue if that fits in with how you work. You may also want to check your GitHub notification settings to see that you do (or don't, if you prefer!) also get email notifications.

Finally, don't hesitate to ask many any questions you might have about the process.
