### 2024-03-07 #6234 @lheckmann
> Hi @rrjbca, @lheckmann,
>
> I see you've started your reviews and was wondering if there was anything you already wanted to comment on. JOSS reviews are intended to be iterative and you're encouraged to comment on whatever parts of the submission you've reviewed as the proceed. That way, the authors can start working on those elements (e.g. installation) while you continue to review others (e.g. testing and functionality).
>
> Let me know if you have any questions!

Sorry for the late reply! There is one thing already:

For the installation, there are instructions for pip but not for conda, which would be beneficial to not cause inconsistencies with other installed python packages.
How do I best communicate this to the authors?
