### 2024-03-10 #6234 @athob
> Sorry for the late reply! There is one thing already:
>
> For the installation, there are instructions for pip but not for conda, which would be beneficial to not cause inconsistencies with other installed python packages. How do I best communicate this to the authors?

Thanks for your feedback @lheckmann, I assume I should be able to respond to this here.

Currently the package is not published on the PyPI repository, so the preferred pip installation instructions point to the GitHub repository where the package source code is hosted. I am not aware of a way to implement something similar with conda, is there one?
