### 2024-03-11 #6234 @lheckmann
> > Sorry for the late reply! There is one thing already:
> > For the installation, there are instructions for pip but not for conda, which would be beneficial to not cause inconsistencies with other installed python packages. How do I best communicate this to the authors?
>
> Thanks for your feedback @lheckmann, I assume I should be able to respond to this here.
>
> Currently the package is not published on the PyPI repository, so the preferred pip installation instructions point to the GitHub repository where the package source code is hosted. I am not aware of a way to implement something similar with conda, is there one?

Okay true,

I now did:
` conda create --name ananke python==3.9`
`conda activate ananke`
and then the pip install command you mentioned to have it separate from my other python installations. But this can then be also left to the user.

The only two other points for the installation:
- jupyter is missing in the installation setup but needed for the tutorial
- For the additional package (numpy etc) no version are specified right? It works with current ones, but this could change in the future so might be good to specify the ones used.
