### 2024-03-12 eaa0626
Update README.md
adding a dedicated instruction for installing Jupyter
