### 2024-03-12 #6234 @athob
Indeed thanks for noticing that.

Regarding jupyter, `py-ananke` itself does not require it to run its modules, so I would prefer to not add it to the package dependencies. To solve that, I have added a couple sentences in the Readme section where the tutorial notebook is mentioned. The added text points at the dedicated Project Jupyter webpage where the collaboration posted instructions for installing jupyter.

Regarding the version specs of the existing dependencies, I am going to better constrain them by imposing the major version of each dependency for both `py-ananke` and its submodules `py-EnBiD-ananke` and `py-Galaxia-ananke`. For example, I am already currently imposing a minimum version for `vaex` in the dependencies of `py-Galaxia-ananke`, but I should also constrain the dependency to no more than versions `4.x` in case a version `5.x` is published with non-backward compatible features. I will send an update here when I'll have added those requirements.
