+++ joss/paper.bib
@@ -1079,6 +1079,15 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{scikit-learn:2011,
+       author = {Pedregosa, F. and Varoquaux, G. and Gramfort, A. and Michel, V. and Thirion, B. and Grisel, O. and Blondel, M. and Prettenhofer, P. and Weiss, R. and Dubourg, V. and Vanderplas, J. and Passos, A. and Cournapeau, D. and Brucher, M. and Perrot, M. and Duchesnay, E.},
+        title = {Scikit-learn: Machine Learning in {P}ython},
+      journal = {Journal of Machine Learning Research},
+       volume = {12},
+        pages = {2825--2830},
+         year = {2011}
+}
+
 @SOFTWARE{pandas:2023,
   author       = {{The pandas development team}},
   title        = {pandas-dev/pandas: Pandas},
