+++ setup.py
@@ -38,7 +38,7 @@ setup(name=NAME,
       python_requires='>=3.7.12',
       packages=[NAME],
       package_dir={'': SRC_DIR},
-      install_requires=["numpy", "scipy", "pandas",
+      install_requires=["numpy>=1.22,<2", "scipy>=1.7.2,<2", "pandas>=2,<3",
                         f"EnBiD_ananke @ file://{(ROOT_DIR / PYENBID).resolve()}",
                         f"Galaxia_ananke @ file://{(ROOT_DIR / PYGALAXIA).resolve()}"]
       )
