### 2024-03-13 #6234 @athob
Thank you for your suggestion @lheckmann, I have tested the main package and its submodules for what would be acceptable minimal version requirements, and added those to the dependencies specifications. I have pushed the updates to the `main` branch of each repository. I also had some minor patching updates that I included in these newer versions.
