### 2024-03-18 #6234 @lheckmann
Perfect, thank you!

I've now started to test the code more properly and have some points/questions to remark:

1. `an.Ananke.make_dummy_particles_input ` is not very well document. I understand it is just to try the code, but help function only explains `n_part` as an input parameter, even though `with_densities=False` is stated as a parameter as well.
2. Why is the observer position chosen randomly in the example? Wouldn't it make most sense to place it at the GAIA position?
3. Is there a way to check the Input objects and its methods mentioned in L150 in the pdf? Or should they only operate in the background?
4. Same for the py-anake classes mentioned afterwards
5. I would advice on improving the docs for the `ananke.run()` function since it runs the main pipeline. In the pdf the different steps are indicated, but in the code doc it just refers to running the pipeline. Also, are there test implemented checking the different substeps?

That is it for now, I'll continue with checking the final output tables etc. Just wanted to leave the intermediate comments here.
