+++ src/ananke/__init__.py
@@ -16,7 +16,20 @@ from .__metadata__ import *
 from .constants import *
 from .Ananke import Ananke
 
-__all__ = ['Ananke']
+__all__ = ['Ananke', 'make_dummy_particles_input', 'display_available_photometric_systems', 'display_density_docs', 'display_EnBiD_docs', 'display_extinction_docs', 'display_errormodel_docs']
+
+
+make_dummy_particles_input = Ananke.make_dummy_particles_input
+
+display_available_photometric_systems = Ananke.display_available_photometric_systems
+
+display_density_docs = Ananke.display_density_docs
+
+display_EnBiD_docs = Ananke.display_EnBiD_docs
+
+display_extinction_docs = Ananke.display_extinction_docs
+
+display_errormodel_docs = Ananke.display_errormodel_docs
 
 
 if __name__ == '__main__':
