### 2024-03-21 c77e081
Adding support for user to provide observer's velocity & adapting to latest dev version of Galaxia
