+++ src/ananke/__metadata__.py
@@ -8,7 +8,7 @@ __author__ = "Adrien Thob"
 __copyright__ = None
 __credits__ = ["Adrien Thob", "Robyn Sanderson", "Andrew Eden"]
 __license__ = "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)"
-__version__ = "0.1.0b3.dev1"
+__version__ = "0.1.1.dev0"
 __maintainer__ = "Adrien Thob"
 __email__ = "athob@sas.upenn.edu"
 __url__ = "https://github.com/athob/py-ananke"
