### 2024-03-21 a159325
Patching observer velocity implementation
