### 2024-03-22 6e80a74
Fixing new Observer implementation
