### 2024-03-22 3ddf12b
Aligning python version requirement with numpy minimal requirement
