### 2024-03-22 abc8a18
Need to update notebook to reflect recent documentation changes
