### 2024-03-22 #6234 @athob
Thanks for your feedback, I have pushed some updates in the last couple of days to address your comments. Note that those updates are currently available in the `develop` branch (at the time of this message under [this commit state](https://github.com/athob/py-ananke/tree/abc8a18681c8fd0302b5809e5293b0c66c6c55c3)):

1. I have improved the documentation for this method, including the description of the `with_densities` kwarg, but also adding a description of what columns are contained in the output dictionary similar to that shown in the notebook. Moreover, given that this method is a class method of the class `Ananke`, I have made it possible to call it directly from the package namespace `ananke` (along with a few other such classmethods for which it would streamline usage to not reference the `Ananke` class namespace).
2. Most of the data shown in the notebook example is completely random by choice, as this allows me to not rely on providing the data of an existing simulation with this notebook. That said, I am hoping in the future to have a notebook example with such realistic galaxy input data downloaded from a publicly available simulation with a small memory footprint to keep it portable. Your comment also reminded me of my plans to make the observer's position default on that of our solar system with respect to our galaxy, which I have now implemented. To add up, I completed the observer coordinates customization feature by also allowing the user to provide velocity coordinates, with the default set on the Sun velocity in our galaxy.
3. It is possible to see the implementation of the `Input` class in the [corresponding py-Galaxia-ananke submodule source file](https://github.com/athob/py-Galaxia-ananke/blob/develop/src/Galaxia_ananke/Input.py). Note however that objects of this class, and of all those available in the py-Galaxia-ananke in general, should not be directly constructed by an `ananke` user. The only such objects that an `ananke` user may end up interacting with are objects of the [`Output` class](https://github.com/athob/py-Galaxia-ananke/blob/develop/src/Galaxia_ananke/Output.py) which I attempted to properly document.
4. Regarding the classes that are in the `ananke` package, only `Ananke` objects should be manipulated by an `ananke` user. The others are for internal use.
5. I will be working on improving the docs for the `Ananke.run` method. Regarding tests, I had started to build a suite of tests for the 3 packages (py-ananke and its 2 submodules) starting with submodule py-EnBiD-ananke which has a relatively complete set. I am still working on building the suite of tests for py-Galaxia-ananke. Note that since the major contributing feature of these 2 submodules is to wrap over existing C++ software, my main focus has been to design their tests to evaluate these wrapping functionalities, trusting that the backend C++ software is operational and accurate. I intend eventually to test beyond the wrapping functionalities, in particular when designing the tests for the main py-ananke module which includes other functionalities of its own that don't relate to wrapping C++ software.

I am looking forward to hearing back from you regarding your other comments.
