### 2024-04-04 #6234 @lheckmann
Thank you for all the explanations and adaptions of the code.
I now wanted to try the changes on the develop branch, but when I try to import ananke it gives me the error:
```
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/home/lea/.local/lib/python3.9/site-packages/ananke/__init__.py", line 17, in <module>
    from .Ananke import Ananke
  File "/home/lea/.local/lib/python3.9/site-packages/ananke/Ananke.py", line 15, in <module>
    from .Universe import Universe
  File "/home/lea/.local/lib/python3.9/site-packages/ananke/Universe.py", line 12, in <module>
    from Galaxia_ananke.defaults import DEFAULTS_FOR_PARFILE
ModuleNotFoundError: No module named 'Galaxia_ananke.defaults'
```

what I did, was switching to the origins/develop branch and am on this HEAD:
`HEAD is now at abc8a18 Need to update notebook to reflect recent documentation changes`
and then re-run `pip install .`
