### 2024-04-04 #6234 @athob
Thanks for the feedback and I think the re-installation didn't update the submodule `Galaxia_ananke` properly. Since you ran `pip install .`, I assume that you cloned the repository and are running the command from it. There is a known issue at re-installation between the submodules and the `pip` cache, I haven't found a proper way to address it yet. Could you try to run the command `pip install . --no-cache-dir` instead? If this is successful, and the import proceed without error, you should end up with the following versions:

```
import ananke, Galaxia_ananke, EnBiD_ananke
print(ananke.__version__)          #0.1.1.dev0
print(EnBiD_ananke.__version__)    #0.1.3b2.dev0
print(Galaxia_ananke.__version__)  #0.3.1.dev0
```

My guess is that the first installation stored in the `pip` cache the outdated versions of both submodules `EnBiD_ananke` and `Galaxia_ananke` that it installed in the process. As a result, while the re-installation `pip` should still update the git submodules to their correct versions, it also ignores the newer versions all together and instead use the cached ones.

Now while typing this, I'm realizing that I should try experimenting forcing the version number of the submodules. I just need to look into how to do it jointly with the way it's done now i.e. specifying their location in the repository since they only exist as submodules at this time.
