### 2024-04-10 #6234 @rrjbca
Hi @athob, first my apologies for the delay in completing my review. In general both the software and manuscript are already at a very high standard and I congratulate you on the excellent work. I have on small specific request to make the API documentation more accessible for users which I have summarized in [Issue #7](https://github.com/athob/py-ananke/issues/7). Once this has been addressed I would be happy to recommend the article for publication.

In addition, while reviewing the software I recognized a few further improvements that I would recommend, but these are beyond the scope of the review criteria and I only offer them as suggestions that you may choose to work on in the future:

- Make tagged releases of py-ananke and distribute via PyPi
- Instead of git submodules, also make tagged releases of py-EnBiD-ananke and py-Galaxia-ananke, distribute them via pypi and add them as dependencies of py-ananke.
- The example jupyter notebook satisfies the review criteria for a manual test to verify functionality, but it would be advisable to adapt it into an automated test.
- A tool like sphinx could be used to create user-friendly  html documentation that incorporates the python docstrings.
