+++ src/ananke/constants.py
@@ -17,4 +17,4 @@ VEL_TAG = 'vel'
 # Rachford et al 2009 (FUSE) measure N_H/E(B-V) (N_H=N_HI+2N_H2) = 5.8e21 H/cm^2/mag
 # many thanks to Julianne Dalcanton for pointing me to this paper!!
 Q_DUST = 4.0e-23 #1.0/5.8e21
-TOTAL_TO_SELECTIVE = 3.1
+TOTAL_TO_SELECTIVE = 3.1  #R_V value
