+++ src/ananke/_default_error_model.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 """
 Contains the definition of default error model estimators for
-existing isochrones of Galaxia.
+existing photometric systems of Galaxia.
 
 Please note that this module is private.
 """
