### 2024-04-17 97e9642
Switching distutils to setuptools and adding implementations to pull submodules metadata
