### 2024-04-17 ded867e
Modified setup.py to detect if pip is running setup.py egg-info and force a call of pip cache remove *_ananke to get rid of existing wheels
Ugly fix, but should solve the issue with pip ignoring updated submodules
