### 2024-04-17 #6234 @athob
Hi @rrjbca, @lheckmann, thanks both for the latest feedback! I will be working on your comments in the coming weeks (I also still need to improve the documentation for the method `Ananke.run` as @lheckmann previously suggested it).

I just wanted to give a small update: I was able to implement a temporary fix to the issue that was ongoing with `pip` keeping wheels in cache of the submodules, and ignoring the updated versions in the process. It's a bit of a ugly fix truly as it calls `pip cache remove` from within the `setup.py` file (and in the future I'll need to think more carefully about it when I'll publish the package on the PyPI), but I wanted to have this done so future updates to the code requires only to run `pip` once.

I'll add more when I'll have addressed your latest comments! Thank for your time
