### 2024-04-18 ed47530
Adapting to latest version of Galaxia_ananke
