+++ joss/paper.bib
@@ -1173,6 +1173,49 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{astroquery:2019,
+       author = {{Ginsburg}, Adam and {Sip{\H{o}}cz}, Brigitta M. and {Brasseur}, C.~E. and {Cowperthwaite}, Philip S. and {Craig}, Matthew W. and {Deil}, Christoph and {Guillochon}, James and {Guzman}, Giannina and {Liedtke}, Simon and {Lian Lim}, Pey and {Lockhart}, Kelly E. and {Mommert}, Michael and {Morris}, Brett M. and {Norman}, Henrik and {Parikh}, Madhura and {Persson}, Magnus V. and {Robitaille}, Thomas P. and {Segovia}, Juan-Carlos and {Singer}, Leo P. and {Tollerud}, Erik J. and {de Val-Borro}, Miguel and {Valtchanov}, Ivan and {Woillez}, Julien and {Astroquery Collaboration} and {a subset of astropy Collaboration}},
+        title = "{astroquery: An Astronomical Web-querying Package in Python}",
+      journal = {\aj},
+     keywords = {astronomical databases: miscellaneous, virtual observatory tools, Astrophysics - Instrumentation and Methods for Astrophysics},
+         year = 2019,
+        month = mar,
+       volume = {157},
+       number = {3},
+          eid = {98},
+        pages = {98},
+          doi = {10.3847/1538-3881/aafc33},
+archivePrefix = {arXiv},
+       eprint = {1901.04520},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019AJ....157...98G},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@MISC{svo_fps:2012,
+       author = {{Rodrigo}, Carlos and {Solano}, Enrique and {Bayo}, Amelia},
+        title = "{SVO Filter Profile Service Version 1.0}",
+ howpublished = {IVOA Working Draft 15 October 2012},
+         year = 2012,
+        month = oct,
+        pages = {1015},
+          doi = {10.5479/ADS/bib/2012ivoa.rept.1015R},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2012ivoa.rept.1015R},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@INPROCEEDINGS{svo_fps:2020,
+       author = {{Rodrigo}, C. and {Solano}, E.},
+        title = "{The SVO Filter Profile Service}",
+    booktitle = {XIV.0 Scientific Meeting (virtual) of the Spanish Astronomical Society},
+         year = 2020,
+        month = jul,
+          eid = {182},
+        pages = {182},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020sea..confE.182R},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
 @ARTICLE{vaex:2018,
        author = {{Breddels}, Maarten A. and {Veljanoski}, Jovan},
         title = "{Vaex: big data exploration in the era of Gaia}",
