### 2024-04-18 1860f7c
Now using universal extinction model from Wang % Chen 2019 as default for extinction coefficients
