### 2024-04-26 3deba3e
Adding zeropoint querying feature & fixing issue with observer coordinates key names
