### 2024-05-14 b7c1f07
Preparing the post-processing pipelines / Adapting the observed magnitude post-processing / Changing the 'expand_and_apply' instance methods in the extinction and error model driver to static methods
