### 2024-05-15 dcb9939
Adapted the extinction driver pipeline the new py-Galaxia output postprocessing framework
