+++ src/ananke/ExtinctionDriver.py
@@ -6,12 +6,15 @@ Please note that this module is private. The ExtinctionDriver class is
 available in the main ``ananke`` namespace - use that instead.
 """
 from __future__ import annotations
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Union, Set, List, Dict, Callable
+from numpy.typing import ArrayLike, NDArray
 from warnings import warn
+from functools import cached_property
 from collections.abc import Iterable
 import numpy as np
 import scipy as sp
-import scipy.interpolate  # needed for python==3.7
+from scipy.interpolate import LinearNDInterpolator
+import pandas as pd
 
 from . import utils
 from ._default_extinction_coeff import *
@@ -19,6 +22,7 @@ from .constants import *
 
 if TYPE_CHECKING:
     from .Ananke import Ananke
+    import Galaxia_ananke as Galaxia
 
 __all__ = ['ExtinctionDriver']
 
@@ -69,8 +73,27 @@ class ExtinctionDriver:
         self._test_extinction_coeff()
     
     __init__.__doc__ = __init__.__doc__.format(Q_DUST=Q_DUST, TOTAL_TO_SELECTIVE=TOTAL_TO_SELECTIVE)
+
+    def __getattr__(self, item):
+        if (item in self.ananke.__dir__() and item.startswith('particle')):
+            return getattr(self.ananke, item)
+        else:
+            return self.__getattribute__(item)
+
+    @property
+    def ananke(self) -> Ananke:
+        return self.__ananke
+
+    @property
+    def particle_column_densities(self) -> NDArray:
+        return self.particles[self._col_density] if self._col_density in self.particles else np.nan*self.particle_masses
     
-    def _make_interpolator(self):
+    @property
+    def galaxia_output(self) -> Galaxia.Output:
+        return self.ananke._galaxia_output
+    
+    @cached_property
+    def column_density_interpolator(self) -> LinearNDInterpolator:
         # center particle coordinates on the observer
         xhel_p = self.ananke.particle_positions - self.ananke.observer_position[:3]
         # TODO coordinates.SkyCoord(**dict(zip([*'uvw'], xhel_p.T)), unit='kpc', representation_type='cartesian', frame='galactic') ?
@@ -81,65 +104,18 @@ class ExtinctionDriver:
         # create a mask for the particles that are within the shell
         sel_interp = (rmin<dhel_p) & (dhel_p<rmax)
         # get the array of column densities input by the user to each particle
-        lognh = self.column_densities
+        lognh = self.particle_column_densities
         # generate the interpolator to use to get the column densities at positions in and around the particles
-        self.__interpolator = sp.interpolate.LinearNDInterpolator(xhel_p[sel_interp],lognh[sel_interp],rescale=False)  # TODO investigate NaN outputs from interpolator
+        self.__interpolator = LinearNDInterpolator(xhel_p[sel_interp],lognh[sel_interp],rescale=False)  # TODO investigate NaN outputs from interpolator
         return self.__interpolator
 
-    def __getattr__(self, item):
-        if (item in self.ananke.__dir__() and item.startswith('particle')):
-            return getattr(self.ananke, item)
-        else:
-            return self.__getattribute__(item)
-
-    @property
-    def ananke(self):
-        return self.__ananke
-
-    @property
-    def column_densities(self):
-        return self.particles[self._col_density] if self._col_density in self.particles else np.nan*self.particle_masses
-    
-    @property
-    def galaxia_output(self):
-        return self.ananke._galaxia_output
-    
-    @property
-    def galaxia_pos(self):
-        return np.array(self.galaxia_output[self._galaxia_pos])
-    
-    @property
-    def column_density_interpolator(self):
-        if self.__interpolator is None:
-            return self._make_interpolator()
-        else:
[…]
