### 2024-05-15 292f9fa
Adapted the error model driver pipeline to the new py-Galaxia output postprocessing framework
