+++ src/ananke/ErrorModelDriver.py
@@ -6,10 +6,12 @@ Please note that this module is private. The ErrorModelDriver class is
 available in the main ``ananke`` namespace - use that instead.
 """
 from __future__ import annotations
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Union, Set, List, Dict, Callable
+from numpy.typing import ArrayLike, NDArray
 from warnings import warn
 from collections.abc import Iterable
 import numpy as np
+import pandas as pd
 
 from . import utils
 from ._default_error_model import *
@@ -17,6 +19,7 @@ from .constants import *
 
 if TYPE_CHECKING:
     from .Ananke import Ananke
+    import Galaxia_ananke as Galaxia
 
 __all__ = ['ErrorModelDriver']
 
@@ -60,20 +63,20 @@ class ErrorModelDriver:
             return self.__getattribute__(item)
 
     @property
-    def ananke(self):
+    def ananke(self) -> Ananke:
         return self.__ananke
 
     @property
-    def galaxia_output(self):
+    def galaxia_output(self) -> Galaxia.Output:
         return self.ananke._galaxia_output
 
     @staticmethod
-    def _expand_and_apply_error_model(df, error_model):
+    def _expand_and_apply_error_model(df, error_model) -> Dict[str, ArrayLike]:
         if not isinstance(error_model, Iterable):
             error_model = [error_model]
         return {key: error for error_dict in [(err_model(df) if callable(err_model) else err_model) for err_model in error_model] for key,error in error_dict.items()}  # TODO adapt to dataframe type of output?
 
-    def _test_error_model(self):
+    def _test_error_model(self) -> None:
         dummy_df = utils.RecordingDataFrame([], columns = self.ananke.galaxia_catalogue_keys + self._extra_output_keys)  # TODO make use of dummy_df.record_of_all_used_keys
         dummy_df.loc[0] = np.nan
         try:
@@ -83,42 +86,54 @@ class ErrorModelDriver:
         utils.compare_given_and_required(dummy_err.keys(), set(), self.ananke.galaxia_catalogue_mag_and_astrometrics, error_message="Given error model function returns wrong set of keys")
     
     @property
-    def _sigma_keys(self):
+    def _sigma_keys(self) -> Set[str]:
         return set(map(self._sigma_template, self.ananke.galaxia_catalogue_mag_names))
 
     @property
-    def _error_keys(self):
+    def _error_keys(self) -> Set[str]:
         return set(map(self._error_template, self.ananke.galaxia_catalogue_mag_names))
+
+    @classmethod
+    def __pp_pipeline(cls, df: pd.DataFrame, error_keys: Set[str],
+                           error_model: List[Union[Callable[[pd.DataFrame],
+                                                            Dict[str, NDArray]],
+                                                   Dict[str, float]]]) -> None:
+        if error_keys.difference(df.columns):
+            for prop_name, error in cls._expand_and_apply_error_model(df, error_model).items():
+                # pre-generate the keys to use for the standard error and its actual gaussian drawn error of property prop_name
+                prop_sig_name, prop_err_name = cls._sigma_template(prop_name), cls._error_template(prop_name)
+                # assign the column of the standard error values for property prop_name in the final catalogue output 
+                df[prop_sig_name] = error
+                # assign the column of the actual gaussian drawn error values for property prop_name in the final catalogue output 
+                df[prop_err_name] = error*np.random.randn(df.shape[0])
+                # add the drawn error value to the existing quantity for property prop_name
+                df[prop_name] += df[prop_err_name]
+                # with_columns.append(prop_name)
 
     @property
     def errors(self):
-        if self._error_keys.difference(self.galaxia_output.columns):
[…]
