+++ src/ananke/Ananke.py
@@ -5,12 +5,15 @@ Contains the Ananke class definition
 Please note that this module is private. The Ananke class is
 available in the main ``ananke`` namespace - use that instead.
 """
+from typing import TYPE_CHECKING, Any, Union, Tuple, List, Dict
+from numpy.typing import NDArray
 from warnings import warn
 import re
 import numpy as np
 import pandas as pd
 
 import Galaxia_ananke as Galaxia
+import Galaxia_ananke.photometry as Galaxia_photo
 
 from .constants import *
 from .Universe import Universe
@@ -18,6 +21,10 @@ from .Observer import Observer
 from .DensitiesDriver import DensitiesDriver
 from .ExtinctionDriver import ExtinctionDriver
 from .ErrorModelDriver import ErrorModelDriver
+
+if TYPE_CHECKING:
+    from Galaxia_ananke.photometry.PhotoSystem import PhotoSystem
+    from astropy.units import Quantity
 
 __all__ = ['Ananke']
 
@@ -51,7 +58,8 @@ class Ananke:
     _intrinsic_mag_formatter = '{}_Intrinsic'
     _intrinsic_mag_template = _intrinsic_mag_formatter.format
 
-    def __init__(self, particles, name, ngb=64, d_params={}, e_params={}, err_params={}, **kwargs) -> None:
+    def __init__(self, particles: Dict[str, NDArray], name: str, ngb: int = 64,
+                 d_params: Dict[str, Any] = {}, e_params: Dict[str, Any] = {}, err_params: Dict[str, Any] = {}, **kwargs: Dict[str, Any]) -> None:
         """
             Parameters
             ----------
@@ -149,19 +157,19 @@ class Ananke:
             make_dummy_particles_input to generate a dummy example of such input
             dictionary.
         """
-        self.__particles = particles
-        self.__name = name
-        self.__ngb = ngb
-        self.__universe_proxy = self._prepare_universe_proxy(kwargs)
-        self.__photo_sys = kwargs.pop(self._photo_sys, Galaxia.DEFAULT_PSYS)
-        self.__observer_proxy = self._prepare_observer_proxy(kwargs)
-        self.__parameters = kwargs
-        self.__densitiesdriver_proxy = self._prepare_densitiesdriver_proxy(d_params)
-        self.__extinctiondriver_proxy = self._prepare_extinctiondriver_proxy(e_params)
-        self.__errormodeldriver_proxy = self._prepare_errormodeldriver_proxy(err_params)
-        self.__galaxia_input = None
-        self.__galaxia_survey = None
-        self.__galaxia_output = None
+        self.__particles: Dict[str, NDArray] = particles
+        self.__name: str = name
+        self.__ngb: int = ngb
+        self.__universe_proxy: Universe = self._prepare_universe_proxy(kwargs)
+        self.__photo_sys: str = kwargs.pop(self._photo_sys, Galaxia.DEFAULT_PSYS)
+        self.__observer_proxy: Observer = self._prepare_observer_proxy(kwargs)
+        self.__parameters: Dict[str, Any] = kwargs
+        self.__densitiesdriver_proxy: DensitiesDriver = self._prepare_densitiesdriver_proxy(d_params)
+        self.__extinctiondriver_proxy: ExtinctionDriver = self._prepare_extinctiondriver_proxy(e_params)
+        self.__errormodeldriver_proxy: ErrorModelDriver = self._prepare_errormodeldriver_proxy(err_params)
+        self.__galaxia_input: Union[Galaxia.Input, None] = None
+        self.__galaxia_survey: Union[Galaxia.Survey, None] = None
+        self.__galaxia_output: Union[Galaxia.Output, None] = None
 
[…]

[…]
