### 2024-05-16 c81e432
Adding support print when post-processing pipelines are being ran
