### 2024-05-16 21a4c3a
Investigating threading issue - using https://github.com/scipy/scipy/issues/8856#issuecomment-390917039
