+++ src/ananke/Ananke.py
@@ -8,6 +8,7 @@ available in the main ``ananke`` namespace - use that instead.
 from typing import TYPE_CHECKING, Any, Union, Tuple, List, Dict
 from numpy.typing import NDArray
 from Galaxia_ananke.photometry.PhotoSystem import PhotoSystem
+from astropy.units import Quantity
 from warnings import warn
 import re
 import numpy as np
@@ -23,8 +24,6 @@ from .DensitiesDriver import DensitiesDriver
 from .ExtinctionDriver import ExtinctionDriver
 from .ErrorModelDriver import ErrorModelDriver
 
-if TYPE_CHECKING:
-    from astropy.units import Quantity
 
 __all__ = ['Ananke']
 
