### 2024-05-17 c6367d8
Improving metadata implementation
