+++ src/ananke/__metadata__.py
@@ -2,14 +2,143 @@
 """
 Contains the ananke module metadata.
 """
-__all__ = ['__author__', '__copyright__', '__credits__', '__license__', '__version__', '__maintainer__', '__email__', '__url__', '__status__']
+import _sitebuiltins
+import textwrap
+from datetime import date
 
+__all__ = ['__project__', '__description__', '__url__', '__year__', '__author__', '__contributors__', '__license__',
+           '__version__', '__date__', '__maintainer__', '__email__', '__classifiers__',
+           '__credits__', '__doi__', '__citation__', '__adsurl__', '__bibtex__', '__copyright__', '__readme__',
+           'copyright', 'credits', 'license', 'readme']
+
+# PROJECT METADATA
+__project__ = "py-ananke"
+__description__ = "Provides a pipeline to generate mock synthetic star catalogs from star particle based simulated data."
+__url__ = "https://github.com/athob/py-ananke"
+__year__ = 2022
 __author__ = "Adrien Thob"
-__copyright__ = None
-__credits__ = ["Adrien Thob", "Robyn Sanderson", "Andrew Eden"]
-__license__ = "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)"
+__contributors__ = ["Robyn Sanderson", "Andrew Eden"]
+__supporters__ = ["Farnik Nikakhtar", "Nondh Panithanpaisal", "Nicolas Garavito-Camargo", "Sanjib Sharma"]
+__feedback__ = ["the extended Galaxy Dynamics @ UPenn group", "the participants of the \"anankethon\" workshops"]
+__license__ = "GNU General Public License v3 or later (GPLv3+)"  # TODO change to longer version?
+__license_classifier__ = "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)"
+__license_short__ = "Licensed under the GNU GPL v3 or later"
+
+# SOFTWARE METADATA
 __version__ = "0.1.1.dev3"
+__date__: date = date(2024, 5, 14)  # TODO how to automatize based on commit day?
 __maintainer__ = "Adrien Thob"
 __email__ = "athob@sas.upenn.edu"
-__url__ = "https://github.com/athob/py-ananke"
-__status__ = "Development Status :: 4 - Beta"
+__status_classifier__ = "Development Status :: 4 - Beta"
+__classifiers__ = [
+    __status_classifier__,
+    "Environment :: Console",
+    "Intended Audience :: Science/Research",
+    __license_classifier__,
+    "Natural Language :: English",
+    "Operating System :: Unix",
+    "Programming Language :: Python :: 3",
+    "Topic :: Scientific/Engineering :: Astronomy"
+    ],
+
+# CREDITS METADATA
+__credits__ = textwrap.fill(f"""
+Thanks to {', '.join(__contributors__)[::-1].replace(',','& ', 1)[::-1]} for their contributions,
+to {', '.join(__supporters__)[::-1].replace(',','& ', 1)[::-1]} for their support & guidance,
+and to the wider community for their suggestions and feedback,
+including {', '.join(__feedback__)[::-1].replace(',','& ', 1)[::-1]}.
+""".strip('\n'))
+
+# CITING METADATA
+__doi__ = "10.48550/arXiv.2312.02268"
+__citation__ = textwrap.fill("""
+Thob, Adrien C. R. et al. 2023, “Generating synthetic star catalogs
+from simulated data for next-gen observatories with py-ananke”,
+arXiv e-prints, doi:10.48550/arXiv.2312.02268.
+""".strip('\n'))
[…]
