+++ setup.py
@@ -18,22 +18,13 @@ setup(name=NAME,
       author_email=__email__,
       maintainer=__maintainer__,
       maintainer_email=__email__,
-      url="https://github.com/athob/py-ananke",
-      description="",
+      url=__url__,
+      description=f"{__project__}: {__description__}",
       long_description=long_description,
       long_description_content_type="text/markdown",
-      classifiers=[
-          __status__,
-          "Environment :: Console",
-          "Intended Audience :: Science/Research",
-          __license__,
-          "Natural Language :: English",
-          "Operating System :: Unix",
-          "Programming Language :: Python :: 3",
-          "Topic :: Database",
-          "Topic :: Scientific/Engineering :: Astronomy",
-          "Topic :: Software Development :: Version Control :: Git"
-      ],
+      classifiers=__classifiers__,
+      license=__license__,
+      copyright=__copyright__,
       python_requires='>=3.8',
       packages=[NAME],
       package_dir={'': SRC_DIR},
