### 2024-05-17 4663f58
Include full license text in metadata & update dev of submodules
