+++ setup.py
@@ -8,7 +8,12 @@ from src.__metadata__ import *
 
 ROOT_DIR = pathlib.Path(__file__).parent
 
+for_all_files = ('__license__', )
+
 long_description = ""
+
+package_data = {NAME: all_files(*for_all_files,
+                                basedir=pathlib.Path(SRC_DIR, NAME))}
 
 EnBiD_meta, Galaxia_meta = check_submodules(ROOT_DIR)
 
@@ -28,6 +33,8 @@ setup(name=NAME,
       python_requires='>=3.8',
       packages=[NAME],
       package_dir={'': SRC_DIR},
+      package_data=package_data,
+      include_package_data=True,
       install_requires=["numpy>=1.22,<2", "scipy>=1.7.2,<2", "pandas>=2,<3",
                         f"EnBiD_ananke @ file://{(ROOT_DIR / PYENBID).resolve()}",
                         f"Galaxia_ananke @ file://{(ROOT_DIR / PYGALAXIA).resolve()}"]
