### 2024-08-08 #6234 @athob
Hi @warrickball, I have been mostly learning to use sphinx and readthedocs lately to build a documentation that would meet the recommendations brought by @rrjbca. I am having some challenges with it and have been working on finding out how to adapt it properly for the various packages. It notably seems to be due to some of the classes methods using a classproperty decorator I took from the astropy package, which the sphinx autosummary struggles with. I understand this review is taking longer than usual, please know that it is currently on top priority on my list of tasks.
