+++ docs/source/reference.rst
@@ -1,0 +1,9 @@
+API reference
+=============
+
+.. autosummary::
+	:toctree: _autosummary
+	:template: custom-module-template.rst
+	:recursive:
+	
+	ananke
