+++ docs/source/index.rst
@@ -1,0 +1,15 @@
+.. py-ananke documentation master file, created by
+   sphinx-quickstart on Wed Aug  7 07:55:17 2024.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+py-ananke documentation
+=======================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+   README.md
+   testing_ananke.ipynb
+   reference.rst
