+++ docs/source/requirements.txt
@@ -1,0 +1,3 @@
+sphinx-copybutton
+myst-nb
+furo
