+++ joss/paper.bib
@@ -1570,6 +1570,21 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{swiftsimio:2020,
+       author = {{Borrow}, Josh and {Borrisov}, Alexei},
+        title = "{swiftsimio: A Python library for reading SWIFT data}",
+      journal = {The Journal of Open Source Software},
+     keywords = {Python, cosmology, i/o, simulations, astronomy},
+         year = 2020,
+        month = aug,
+       volume = {5},
+       number = {52},
+          eid = {2430},
+        pages = {2430},
+          doi = {10.21105/joss.02430},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020JOSS....5.2430B},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
 
 
 #################################
