+++ src/ananke/__metadata__.py
@@ -4,6 +4,7 @@ Contains the ananke module metadata.
 """
 import _sitebuiltins
 import textwrap
+from typing import Final
 from datetime import date
 
 try:
@@ -30,7 +31,7 @@ __license_classifier__ = "License :: OSI Approved :: GNU General Public License 
 __license_short__ = "Licensed under the GNU GPL v3 or later"
 
 # SOFTWARE METADATA
-__version__ = "0.1.1.dev4"
+__version__: Final[str] = "0.1.1.dev4"
 __date__: date = date(2024, 5, 17)  # TODO how to automatize based on commit day?
 __maintainer__ = "Adrien Thob"
 __email__ = "athob@sas.upenn.edu"
