### 2024-08-15 5c0bbf2
Improving sphinx usage with apidoc
