+++ docs/source/conf.py
@@ -3,6 +3,8 @@
 # For the full list of built-in configuration values, see the documentation:
 # https://www.sphinx-doc.org/en/master/usage/configuration.html
 
+package_name = 'ananke'
+
 import os
 import sys
 sys.path.insert(0, os.path.abspath('../..'))
@@ -24,7 +26,6 @@ extensions = [
 	'sphinx.ext.duration',
 	'sphinx.ext.doctest',
 	'sphinx.ext.autodoc',
-	'sphinx.ext.autosummary',
 	'sphinx.ext.inheritance_diagram',
 	'sphinx.ext.napoleon',
 	'sphinx.ext.viewcode',
@@ -33,7 +34,6 @@ extensions = [
     'myst_nb',
 ]
 
-templates_path = ['_templates']
 exclude_patterns = ['.ipynb_checkpoints/*']
 
 
@@ -41,4 +41,31 @@ exclude_patterns = ['.ipynb_checkpoints/*']
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
 html_theme = 'furo'
-html_static_path = ['_static']
+
+# -- Options for automatic API doc
+
+autodoc_default_options = {
+    'member-order': 'bysource',
+    'special-members': '__init__',
+    'ignore-module-all': True
+}
+
+def run_apidoc(_):
+    try:
+        from sphinx.ext.apidoc import main
+    except ImportError:
+        from sphinx.apidoc import main
+
+    sys.path.append(os.path.join(os.path.dirname(__file__), ".."))
+    cur_dir = os.path.abspath(os.path.dirname(__file__))
+
+    api_doc_dir = os.path.join(cur_dir, "modules")
+    module = os.path.join(cur_dir, "../..", f"src/{package_name}")
+    ignore = [
+    ]
+
+    main(["-M", "-f", "-e", "-T", "-d 0", "-o", api_doc_dir, module, *ignore])
+
+
+def setup(app):
+    app.connect("builder-inited", run_apidoc)
