+++ src/constants.py
@@ -1,1 +1,0 @@
-ananke/constants.py
