### 2024-08-16 6c78459
Refactoring submodules that should be private for documentation
