### 2024-08-17 athob/py-ananke#7 @athob
Thank you for your feedback. In order to make use of the existing set of docstrings, and as a starting point to put together an appropriate documentation, I've chosen to use sphinx and readthedocs. The [latest release](../tree/8769c5d2a8b565ada26a6485b660969b6765c270) has the corresponding documentation source incorporated in the repository in the `docs` directory. I also added the readthedocs link in the "About" section for ease of access.

With that done, I will be including each of the items you mentioned and comment back once it's all done.
