### 2024-08-26 c3ae816
Improving automation of doc generation to take into account the README internal links
