### 2024-08-26 e52028e
Improving automation of doc generation to take into account the README internal links
