+++ docs/source/requirements.txt
@@ -1,3 +1,5 @@
 sphinx-copybutton
 myst-nb
 furo
+markdown
+lxml
