### 2024-08-27 8daccd1
Fixing latex dollar maths & improving example notebook documentation & temporarily changing sphinx theme
