### 2024-08-27 athob/py-ananke#7 @athob
I continued working on this issue in the dedicated [7-joss-review-api-doc](../tree/7-joss-review-api-doc) branch. I have also activated the readthedocs autobuild for that branch specifically. This documentation is available at this link https://py-ananke.readthedocs.io/en/7-joss-review-api-doc/.

Lately, I have improved the sphinx workflow so that it can make better use of the existing README on the main github page. An issue I notably had before was related to my use of internal links. It was notably problematic as the example notebook was referred in the README via such internal link, which didn't render well in the sphinx-generated readthedocs.

Since I wanted that example notebook to be the main resource for user to first get their hand on the software, I aimed to include the points you mentioned in it. After fixing the sphinx generation, and the internal link rendering, I improved the existing notebook to reflect those points. I have yet to list the additional `kwargs` that can be passed to `Ananke.run`, I'll comment once that is done.

Note that I am having some other rendering issues for which I have not found a solution yet. It notably relates to the formatting of the output cells, which I haven't a way to make scrollable (for readability), whether in the case of pure text outputs, or html-formatted vaex tables.

I also noticed that the [furo theme](https://sphinx-themes.org/sample-sites/furo/) I was using had poor choice of color in dark mode for the latter tables: for this, I temporarily switched to the default readthedocs theme until I can find a way to use furo.
