+++ docs/source/requirements.txt
@@ -1,3 +1,4 @@
+sphinx-math-dollar
 sphinx-copybutton
 myst-nb
 furo
