### 2024-08-28 e526e67
Improving documentation of method by making it dynamically parse into Galaxia's docstring
