+++ src/ananke/Ananke.py
@@ -17,6 +17,7 @@ import pandas as pd
 import Galaxia_ananke as Galaxia
 import Galaxia_ananke.photometry as Galaxia_photo
 
+from . import utils
 from ._constants import *
 from .Universe import Universe
 from .Observer import Observer
@@ -53,7 +54,6 @@ class Ananke:
     _def_uni_rshell = Universe._default_rshell
     _def_photo_sys = Galaxia.DEFAULT_PSYS
     _def_cmd_mags = Galaxia.DEFAULT_CMD
-    _def_cmd_box = Galaxia.DEFAULT_CMD_BOX
     _intrinsic_mag_formatter = '{}_Intrinsic'
     _intrinsic_mag_template = _intrinsic_mag_formatter.format
 
@@ -74,18 +74,18 @@ class Ananke:
 
             d_params : dict
                 Parameters to configure the kernel density estimation. Use
-                class method display_density_docs to find what parameters can
-                be defined
+                class method ``display_density_docs`` to find what parameters
+                can be defined
 
             e_params : dict
                 Parameters to configure the extinction pipeline. Use class
-                method display_extinction_docs to find what parameters can be
-                defined
+                method ``display_extinction_docs`` to find what parameters can
+                be defined
 
             err_params : dict
                 Parameters to configure the error model pipeline. Use class
-                method display_errormodel_docs to find what parameters can be
-                defined
+                method ``display_errormodel_docs`` to find what parameters can
+                be defined
 
             observer : array-like shape (3,) or dict of array-like shape (3,)
                 Coordinates for the observer in phase space. Position and
@@ -94,67 +94,58 @@ class Ananke:
                 enough. If specifying both position and velocity, please
                 provide a dictionary containing both coordinates as array-like
                 objects of shape (3,), respectively denoting the position and
-                velocity coordinates with keys `{_pos}` and `{_vel}`.
-                Position coordinates default to
-                {_def_obs_position}
-                and velocity coordinates default to
-                {_def_obs_velocity}
+                velocity coordinates with keys ``{_pos}`` and ``{_vel}``.
+                Position coordinates default to::
+                
+                    {_def_obs_position}
+                
+                and velocity coordinates default to::
+                
+                    {_def_obs_velocity}
             
             rshell : array-like shape (2,)
                 Range in kpc of distances from the observer position of the
-                particles that are to be considered. Default to
-                {_def_uni_rshell}.
+                particles that are to be considered. Default to::
+                
+                    {_def_uni_rshell}
 
             photo_sys : string or list
                 Name(s) of the photometric system(s) Galaxia should use to
-                generate the survey. Default to {_def_photo_sys}.
+                generate the survey. Default to ``{_def_photo_sys}``.
                 Available photometric systems can be queried with the class
-                method display_available_photometric_systems.
+                method ``display_available_photometric_systems``.
 
             cmd_magnames : string or dict
                 Names of the filters Galaxia should use for the color-magnitude
                 diagram box selection.
                 The input can be given as string in which case it must meet the
-                following format:
+                following format::
+
                     "band1,band2-band3"
-                where band1 is the magnitude filter and (band2, band3) are the
-                filters that define the band2-band3 color index.
+                
+                where ``band1`` is the magnitude filter and ``(band2, band3)``
+                are the filters that define the ``band2-band3`` color index.
                 Alternatively, a dictionary can be passed with the following
-                format:
+                format::
+
                     dict('magnitude': band1,
                          'color_minuend': band2,
[…]
