+++ docs/source/conf.py
@@ -40,6 +40,7 @@ extensions = [
 
 exclude_patterns = ['.ipynb_checkpoints/*']
 
+napoleon_use_param = False
 
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
