+++ docs/source/conf.py
@@ -42,6 +42,8 @@ exclude_patterns = ['.ipynb_checkpoints/*']
 
 napoleon_use_param = False
 
+nb_execution_timeout = 1000
+
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
