### 2024-08-28 a071902
More formatting improvements & temporarily adjusting nb_st timeout parameter
