+++ docs/source/conf.py
@@ -42,7 +42,7 @@ exclude_patterns = ['.ipynb_checkpoints/*']
 
 napoleon_use_param = False
 
-nb_execution_timeout = 1000
+nb_execution_timeout = 100
 
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
