### 2024-08-28 3a340cb
Improved description of run method in notebook
