### 2024-08-28 817bb3b
Improved description of run method in notebook
