+++ docs/source/conf.py
@@ -42,6 +42,23 @@ exclude_patterns = ['.ipynb_checkpoints/*']
 
 napoleon_use_param = False
 
+myst_enable_extensions = [
+    "amsmath",
+    "attrs_inline",
+    "colon_fence",
+    "deflist",
+    "dollarmath",
+    "fieldlist",
+    "html_admonition",
+    "html_image",
+    "linkify",
+    "replacements",
+    "smartquotes",
+    "strikethrough",
+    "substitution",
+    "tasklist",
+]
+
 nb_execution_timeout = 100
 
 # -- Options for HTML output -------------------------------------------------
