### 2024-08-28 8c02714
New documentation style & formatting
