### 2024-08-28 2bd5b8d
New documentation style & formatting
