### 2024-08-29 athob/py-ananke#7 @athob
In [7-joss-review-api-doc](../tree/7-joss-review-api-doc)'s last few updates, I was able to improve the `Ananke` docstrings in a way that it now dynamically uses the `Galaxia_ananke` docstrings to carry over specifically selected parameters descriptions. This allowed me to also complete the documentation of the `Ananke.run` method, as well as properly identifying the parameters to list in the example notebook.

I also found a better sphinx theme (the [pydata-sphinx-theme](https://sphinx-themes.org/sample-sites/pydata-sphinx-theme)) that allowed better dark mode support than furo while also enabling the horizontal scrolling I needed for the html-formatted vaex tables.

I believe that with this last updates, this completes all the tasks that were listed.
