### 2024-08-29 #6234 @athob
Hi @rrjbca & @lheckmann, my apologies for the time it took me to resume this review.

I was able to address the [Issue #7](https://github.com/athob/py-ananke/issues/7) raised by @rrjbca regarding the need of an appropriate API documentation. I have already merged the newer version in the [repository develop branch](https://github.com/athob/py-ananke/tree/develop).

As mentioned in my previous message, and in the issue's discussion, I chose to use sphinx and readthedocs to build this documentation following [@rrjbca's recommendation](https://github.com/openjournals/joss-reviews/issues/6234#issuecomment-2047204850). To avoid repeating texts unnecessarily, I built the sphinx workflow in such a way that it utilizes the repository readme as an introduction page, and the notebook as an example usage page. I intend to add more notebooks in the future with more specialized examples. In the process, I've also improved how the docstrings were being generated in the package itself. That allowed me to improve the documentation of the `Ananke.run`  method, a task I had left unfinished that was requested by @lheckmann in [a previous comment](https://github.com/openjournals/joss-reviews/issues/6234#issuecomment-2004529264).

With that, I still need to address the questions/recommendations in [@lheckmann's last comment](https://github.com/openjournals/joss-reviews/issues/6234#issuecomment-2057738810). Beside that, I also noted the remaining [recommendations of @rrjbca](https://github.com/openjournals/joss-reviews/issues/6234#issuecomment-2047204850) (described as beyond the scope of the review), specifically the distribution of the package and its submodules on PyPI, and the possibility of turning the example notebook into a functionality test.
