+++ joss/paper.bib
@@ -1268,6 +1268,15 @@ archivePrefix = {arXiv},
           url = "https://doi.org/10.1007/978-1-4842-2677-3_5"
 }
 
+@MISC{docstring_parser,
+       author = {{Kurczewski}, Marcin et al.},
+        title = {`docstring_parser` v0.16},
+         year = {2024},
+    publisher = {GitHub},
+      journal = {GitHub repository},
+          url = {https://github.com/rr-/docstring_parser}
+}
+
 @ARTICLE{EnBiD:2006,
        author = {{Sharma}, Sanjib and {Steinmetz}, Matthias},
         title = "{Multidimensional density estimation and phase-space structure of dark matter haloes}",
