### 2024-08-29 7318a33
docstring_parser formatting in citation
