### 2024-08-29 c5f6168
docstring_parser formatting in citation
