+++ joss/paper.bib
@@ -1270,7 +1270,7 @@ archivePrefix = {arXiv},
 
 @MISC{docstring_parser,
        author = {{Kurczewski}, Marcin et al.},
-        title = {``docstring_parser`` v0.16},
+        title = {\verb|docstring_parser| v0.16},
          year = {2024},
     publisher = {GitHub},
       journal = {GitHub repository},
