### 2024-08-29 12d1a22
Improving on documentation of Ananke.run by including description of output catalog in notes
