### 2024-08-30 4e63f8c
More detailing to documentation
