### 2024-08-30 #6234 @athob
Thanks again for your patience @lheckmann. I have addressed most of your comments with the latest version on [develop](https://github.com/athob/py-ananke/tree/develop). I had some clarifications to add which I discuss below:

> * The output columns specified in the jupyter for the final data frame are not the same as when using help(survey).

Thank you for pointing that out, hopefully the jupyter is now in a more appropriate shape

> * Referring to the columns in the jupyter: There are a few which are not explained: e.g. `gaiadr2_g_bpmag 	gaiadr2_g_bpmag_Err 	gaiadr2_g_bpmag_Intrinsic 	gaiadr2_g_bpmag_Sig` and similar ones with other filters, and ` satid and smass` are not in the explanation. On the other hand `mini` is in the explanation, but the column does not exist in my output survey object.

There was indeed an issue regarding `satid` and `smass`, which I accordingly addressed. I wrongly put `mini` for what actually was `smass`, and I had completely overlooked the `satid` column. Regarding the `gaiadr2_g_{}` and other similarly formatted columns, I added a few comments to the description with examples to discuss those (as well as in the `Ananke.run` documentation)

> * This could be due to me not understanding a certain aspect of the code, but why are there so many rows with nan on many columns (e.g. ra, dec, mub, vr,...) in the survey dataframe?

This is one comment I need to look into further to properly verify what is going on here, but `nan` values can happen in some cases, notably when the extinction driver try to extrapolate the line-of-sight column densities. In normal cases with actual star particles data generated with an appropriate simulation model, this doesn't cause too many issues, in the case of that jupyter notebook where the data given is purely random, it does generate a significant amount of `nan` values. As mentioned in my last comment, I plan to add a couple other notebooks to the documentation meant to showcase a real case of mock catalog, such as the one featured in [Sanderson et al. 2020](https://ui.adsabs.harvard.edu/abs/2020ApJS..246....6S).

> * The method `survey.flush_extra_columns_to_hdf5` does not have any documentation when using help. Please add a short one, also showing how one could specify and output file.

I have documented this method, as well as another new method that was added with the newer updates since your comment (`apply_post_process_pipeline_and_flush`). Regarding how to specify an output file, due to the way the output is actually memory-mapped to an existing backend HDF5 file, the flushing uses that backend file already. With the current implementation, this file is written to the `output_dir` given to the `run` method which defaults to the current working directory. Note that with newer updates that try to implement the possibility of working with a catalog split across multiple files (for parallelizing purpose), there can be more than one file. The default behavior of the pipeline is to assume only one output file, but I need to give more thoughts into how to properly document this feature.

------------------------------------

I will be addressing in my next comment the remaining [recommendations of @rrjbca](../issues/6234#issuecomment-2047204850).
