+++ CITATION.cff
@@ -1,0 +1,43 @@
+date-released: "2023-dec-01"
+repository-code: "https://github.com/athob/py-ananke"
+message: "If you use this software, please cite it as below."
+doi: "10.48550/arXiv.2312.02268"
+title: "Generating synthetic star catalogs from simulated data for next-gen observatories with py-ananke"
+cff-version: "1.2.0"
+authors:
+  - family-names: "Thob"
+    given-names: "Adrien C. R."
+  - family-names: "Sanderson"
+    given-names: "Robyn E."
+  - family-names: "Eden"
+    given-names: "Andrew P."
+  - family-names: "Nikakhtar"
+    given-names: "Farnik"
+  - family-names: "Panithanpaisal"
+    given-names: "Nondh"
+  - family-names: "Garavito-Camargo"
+    given-names: "Nicolás"
+  - family-names: "Sharma"
+    given-names: "Sanjib"
+preferred-citation:
+  type: "article"
+  doi: "10.48550/arXiv.2312.02268"
+  date-released: "2023-dec-01"
+  journal: "arXiv e-prints"
+  title: "Generating synthetic star catalogs from simulated data for next-gen observatories with py-ananke"
+  start: "arXiv:2312.02268"
+  authors:
+    - family-names: "Thob"
+      given-names: "Adrien C. R."
+    - family-names: "Sanderson"
+      given-names: "Robyn E."
+    - family-names: "Eden"
+      given-names: "Andrew P."
+    - family-names: "Nikakhtar"
+      given-names: "Farnik"
+    - family-names: "Panithanpaisal"
+      given-names: "Nondh"
+    - family-names: "Garavito-Camargo"
+      given-names: "Nicolás"
+    - family-names: "Sharma"
+      given-names: "Sanjib"
