+++ CITATION.cff
@@ -2,23 +2,31 @@ date-released: "2023-dec-01"
 repository-code: "https://github.com/athob/py-ananke"
 message: "If you use this software, please cite it as below."
 doi: "10.48550/arXiv.2312.02268"
+url: "https://arxiv.org/abs/2312.02268"
 title: "Generating synthetic star catalogs from simulated data for next-gen observatories with py-ananke"
 cff-version: "1.2.0"
 authors:
   - family-names: "Thob"
-    given-names: "Adrien C. R."
+    given-names: "Adrien C.R."
+    orcid: "https://orcid.org/0000-0001-7928-1973"
   - family-names: "Sanderson"
     given-names: "Robyn E."
+    orcid: "https://orcid.org/0000-0003-3939-3297"
   - family-names: "Eden"
     given-names: "Andrew P."
+    orcid: "https://orcid.org/0009-0004-2146-4856"
   - family-names: "Nikakhtar"
     given-names: "Farnik"
+    orcid: "https://orcid.org/0000-0002-3641-4366"
   - family-names: "Panithanpaisal"
     given-names: "Nondh"
+    orcid: "https://orcid.org/0000-0001-5214-8822"
   - family-names: "Garavito-Camargo"
     given-names: "Nicolás"
+    orcid: "https://orcid.org/0000-0001-7107-1744"
   - family-names: "Sharma"
     given-names: "Sanjib"
+    orcid: "https://orcid.org/0000-0002-0920-809X"
 preferred-citation:
   type: "article"
   doi: "10.48550/arXiv.2312.02268"
@@ -28,16 +36,23 @@ preferred-citation:
   start: "arXiv:2312.02268"
   authors:
     - family-names: "Thob"
-      given-names: "Adrien C. R."
+      given-names: "Adrien C.R."
+      orcid: "https://orcid.org/0000-0001-7928-1973"
     - family-names: "Sanderson"
       given-names: "Robyn E."
+      orcid: "https://orcid.org/0000-0003-3939-3297"
     - family-names: "Eden"
       given-names: "Andrew P."
+      orcid: "https://orcid.org/0009-0004-2146-4856"
     - family-names: "Nikakhtar"
       given-names: "Farnik"
+      orcid: "https://orcid.org/0000-0002-3641-4366"
     - family-names: "Panithanpaisal"
       given-names: "Nondh"
+      orcid: "https://orcid.org/0000-0001-5214-8822"
     - family-names: "Garavito-Camargo"
       given-names: "Nicolás"
+      orcid: "https://orcid.org/0000-0001-7107-1744"
     - family-names: "Sharma"
       given-names: "Sanjib"
+      orcid: "https://orcid.org/0000-0002-0920-809X"
