### 2024-08-31 8879452
Introducing first functionality test adapted from example notebook and allowing pytest to run in the CI/CD action
