+++ .github/workflows/main.yml
@@ -40,10 +40,10 @@ jobs:
         flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
-    # - name: Test with pytest
-    #   run: |
-    #     conda install pytest
-    #     pytest
+    - name: Test with pytest
+      run: |
+        conda install pytest
+        pytest src
   
   develop-post-install:
     runs-on: ubuntu-latest
