### 2024-09-01 23be758
Refactoring and cleaning up setup.py
