+++ setup.py
@@ -1,21 +1,12 @@
 #!/usr/bin/env python
-import pathlib
 from setuptools import setup
 
 from src._build_utils import *
-from src._constants import NAME, SRC_DIR, PYENBID, PYGALAXIA
+from src._constants import NAME, SRC_DIR
 from src.__metadata__ import *
 
-ROOT_DIR = pathlib.Path(__file__).parent
-
-for_all_files = ('__license__', )
-
-long_description = ""
-
-package_data = {NAME: all_files(*for_all_files,
-                                basedir=pathlib.Path(SRC_DIR, NAME))}
-
-EnBiD_meta, Galaxia_meta = check_submodules(ROOT_DIR)
+with open("README.md") as readme:
+    long_description = readme.read()
 
 setup(name=NAME,
       version=__version__,
@@ -33,9 +24,8 @@ setup(name=NAME,
       python_requires='>=3.8',
       packages=[NAME],
       package_dir={'': SRC_DIR},
-      package_data=package_data,
+      package_data=make_package_data(),
       include_package_data=True,
-      install_requires=["numpy>=1.22,<2", "scipy>=1.7.2,<2", "pandas>=2,<3", "docstring_parser>=0.16,<0.17",
-                        f"EnBiD_ananke @ file://{(ROOT_DIR / PYENBID).resolve()}",
-                        f"Galaxia_ananke @ file://{(ROOT_DIR / PYGALAXIA).resolve()}"]
+      install_requires=append_install_requires_with_submodules(
+          ["numpy>=1.22,<2", "scipy>=1.7.2,<2", "pandas>=2,<3", "docstring_parser>=0.16,<0.17"]),
       )
