### 2024-09-01 0ece5df
Adding versioneer to package
