+++ setup.py
@@ -28,4 +28,5 @@ setup(name=NAME,
       include_package_data=True,
       install_requires=append_install_requires_with_submodules(
           ["numpy>=1.22,<2", "scipy>=1.7.2,<2", "pandas>=2,<3", "docstring_parser>=0.16,<0.17"]),
+      cmdclass=make_cmdclass(),
       )
