### 2024-09-02 #6234 @athob
@rrjbca I have address your remaining recommendations with the [latest version `0.2.0b3`](https://github.com/athob/py-ananke/tree/v0.2.0b3).

The code that lived in the example notebook now also lives in an independent test script that is ran automatically by the CI/CD GitHub action. This is at the moment the only test available, I intend in the future to broaden the set of tests, notably have some that are more targeted to test individual features.

Regarding the distribution on the PyPI, I have taken steps towards that goal. I have notably overhauled the installing workflow so I could better integrate the versioneer software into it (that I discovered thanks to [this tutorial](https://jacobtomlinson.dev/series/creating-an-open-source-python-project-from-scratch/)), which the package is now able to use. I wanted to test out the publication on the PyPI but found out the project name `ananke` [is already taken](https://pypi.org/project/ananke/) by a project that looks abandoned. Instead of using a different name than that of its import name, [I have taken measures](https://github.com/pypi/support/issues/4678) to request for the name to be made available under [PEP 541](https://peps.python.org/pep-0541/). I will proceed with the publication once a decision is made on that request. I also want first to test out the publication of `ananke` on the PyPI before proceeding with publishing to Conda Forge, or before looking into publishing the submodules `EnBiD_ananke` and `Galaxia_ananke`.

--------------------

With that I believe I have addressed all the remaining comments @warrickball. @lheckmann @rrjbca I am looking forward for your next comments, once again I am sorry for the time I took before addressing your recommendations.
