### 2024-09-08 #6234 @rrjbca
@athob @warrickball The new documentation and CI pipeline look great; I am happy to recommend the article for publication.
