### 2024-09-12 #6234 @warrickball
I'm having a look over things but wanted to make a quick technical comment:

> I also want first to test out the publication of ananke on the PyPI before proceeding with publishing to Conda Forge, or before looking into publishing the submodules EnBiD_ananke and Galaxia_ananke.

We won't require you to create a Conda package because it looks like you can install `py-ananke` from `pip`, so Conda users can create a Conda environment and then use `pip`. I also don't think we need to wait for the PyPI registration because by configuring the repo as a package, it can be installed using the GitHub URL, as indicated in the instructions.

I'll also note that I'm not sure you need the submodule trick any more either. Because `py-EnBiD-ananke` and `py-Galaxia-ananke` are themselves packages, I believe you could include them as dependencies like e.g.
```
py-EnBiD-ananke @ git+https://github.com/athob/py-Galaxia-ananke@0aac265
```
though don't quote me on the syntax here! I don't expect this to happen now but you might want to use this in the future to simplify the build process.

By the way, I'm having trouble installing `py-ananke` but I'll open an issue in the repo to resolve that.
