### 2024-09-12 athob/py-ananke#11 @warrickball
Installation failure on Python 3.9.5 (RHEL 8.8, fresh environment)
To test the installation process, I just created a fresh Python 3.9.5 environment and ran
```
$ pip install git+https://github.com/athob/py-ananke@main
```
which eventually errored out with
```
Collecting git+https://github.com/athob/py-ananke@main
  Cloning https://github.com/athob/py-ananke (to revision main) to /tmp/pip-req-build-_v05__rf
  Running command git clone -q https://github.com/athob/py-ananke /tmp/pip-req-build-_v05__rf
  Running command git submodule update --init --recursive -q
Checking submodules, running git... Processing /tmp/pip-req-build-_v05__rf/py-EnBiD-ananke
  DEPRECATION: A future pip version will change local packages to be built in-place without first copying to a temporary directory. We recommend you use --use-feature=in-tree-build to test your packages with this new behavior before it becomes the default.
   pip 21.3 will remove support for this functionality. You can find discussion regarding this at https://github.com/pypa/pip/issues/7555.
     command: /rds/homes/b/ballwh/work/pyananke/bin/python3 -c 'import io, os, sys, setuptools, tokenize; sys.argv[0] = '"'"'/tmp/pip-install-l0t1_g7z/enbid-ananke_032fa6dcedae4d11a75fa62510bcd565/setup.py'"'"'; __file__='"'"'/tmp/pip-install-l0t1_g7z/enbid-ananke_032fa6dcedae4d11a75fa62510bcd565/setup.py'"'"';f = getattr(tokenize, '"'"'open'"'"', open)(__file__) if os.path.exists(__file__) else io.StringIO('"'"'from setuptools import setup; setup()'"'"');code = f.read().replace('"'"'\r\n'"'"', '"'"'\n'"'"');f.close();exec(compile(code, __file__, '"'"'exec'"'"'))' egg_info --egg-base /tmp/pip-pip-egg-info-j8latto8
         cwd: /tmp/pip-install-l0t1_g7z/enbid-ananke_032fa6dcedae4d11a75fa62510bcd565/
    Complete output (7 lines):
    Traceback (most recent call last):
      File "<string>", line 1, in <module>
      File "/tmp/pip-install-l0t1_g7z/enbid-ananke_032fa6dcedae4d11a75fa62510bcd565/setup.py", line 7, in <module>
        from src.__metadata__ import *
      File "/tmp/pip-install-l0t1_g7z/enbid-ananke_032fa6dcedae4d11a75fa62510bcd565/src/__metadata__.py", line 38, in <module>
        __date__: datetime = datetime.fromisoformat(f"{__date__[:-2]}:{__date__[-2:]}")
    TypeError: 'NoneType' object is not subscriptable
WARNING: Discarding file:///tmp/pip-req-build-_v05__rf/py-EnBiD-ananke. Command errored out with exit status 1: python setup.py egg_info Check the logs for full command output.
```

My guess is that this is coming up because the function [`get_versions`](https://github.com/athob/py-EnBiD-ananke/blob/0aac2654abb18312b809bc46b3868133a34b7d81/src/EnBiD_ananke/_version.py#L640-L683) in `py-EnBiD-ananke` (though I think the same function is used in `py-ananke` and `py-Galaxia-ananke`) is returning one of the dictionaries with `date = None`.

This is RHEL 8.8. `pip freeze` returns nothing, presumably because I'm in a fresh virtual environment.
