### 2024-09-30 athob/py-ananke#11 @warrickball
Sorry to push but I was wondering if there'd been any progress on this? I know we're waiting for the JOSS review to finish up but if we can resolve this now it'll speed things up once the reviewers are ready to accept.
