### 2024-09-30 athob/py-ananke#11 @athob
Thanks for pushing, I wasn't able to look into it at this time, but I do want to address it promptly. I do expect being able to this Wednesday, but I'll try looking into it before too. I am not familiar with RHEL 8.8, but I'll see if I can reproduce the error.
