### 2024-10-01 athob/py-ananke#11 @athob
I was able to reproduce the error in a [RHEL 8.10 docker container](https://hub.docker.com/r/redhat/ubi8). After setting up a minimalist python version installation (no conda, just python 3.9 with a pip installer), I ended up with the same error. After doing a few tests, I saw that the pip I got installed was version 20.2.4, so I tried upgrading it to its latest version (24.2 through pip). After completing the upgrade, I was able to pip install `py-ananke` without that error.

I have yet to test any version of pip in between, but could you let me know what version of pip are you running for now?
