### 2024-10-01 athob/py-ananke#11 @warrickball
I was using pip version 21.1.1, I think inherited from our OS installation (possibly including that I loaded Python 3.9.5 via [Environment Modules](https://modules.readthedocs.io/en/latest/)). I tried creating a fresh virtual environment (still using the Python 3.9.5 module), upgrading pip to 24.2 and installing `py-ananke` again and I got to build errors in Enbid and Galaxia. I've appended the full output below but I think the key lines are that they get to are
```
building '' extension
gcc -shared -L<very long list of library paths because of modules> -o build/lib.linux-x86_64-cpython-39/.cpython-39-x86_64-linux-gnu.so
gcc: fatal error: no input files
compilation terminated.
error: command '/rds/bear-apps/2021a/EL8-ice/software/GCCcore/10.3.0/bin/gcc' failed with exit code 1
```
for both. I'm guessing there should be a list of source files in there that for some reason is coming up empty.

<summary>More error output</summary>

```
Building wheels for collected packages: ananke, EnBiD-ananke, Galaxia-ananke
  Building wheel for ananke (pyproject.toml) ...
Checking submodules, running git... done
  Created wheel for ananke: filename=ananke-0.2.0b3-py3-none-any.whl size=56751 sha256=4a26ce03c3486b3c41c19d29d8341fcaa7c2242cad204e5ba92d93ef60d70d37
  Stored in directory: /tmp/pip-ephem-wheel-cache-vm_l7osh/wheels/6d/55/b2/a3c3c77d39b8afff99b8f831ac46f242e6f930412587426a4b
  Building wheel for EnBiD-ananke (pyproject.toml) ...
Downloading Enbid -
Unpacking Enbid|
Compiling Enbid
        Configuring
        Running make/
error
  error: subprocess-exited-with-error

  × Building wheel for EnBiD-ananke (pyproject.toml) did not run successfully.
  │ exit code: 1
  ╰─> [138 lines of output]

      Downloading Enbid
      Unpacking Enbid
      Compiling Enbid
          Configuring
          Running make
       Warning: 'classifiers' should be a list, got type 'tuple'
      /tmp/pip-build-env-c6wiibpi/overlay/lib/python3.9/site-packages/setuptools/_distutils/dist.py:261: UserWarning: Unknown distribution option: 'copyright'
        warnings.warn(msg)
      running bdist_wheel
      running build
      running build_py
      creating build/lib.linux-x86_64-cpython-39/EnBiD_ananke
      copying src/EnBiD_ananke/__init__.py -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke
      copying src/EnBiD_ananke/__metadata__.py -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke
      copying src/EnBiD_ananke/_builtin_utils.py -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke
      copying src/EnBiD_ananke/_constants.py -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke
      copying src/EnBiD_ananke/_defaults.py -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke
      copying src/EnBiD_ananke/_templates.py -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke
      copying src/EnBiD_ananke/_version.py -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke
      copying src/EnBiD_ananke/utils.py -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke
      running egg_info
      writing src/EnBiD_ananke.egg-info/PKG-INFO
      writing dependency_links to src/EnBiD_ananke.egg-info/dependency_links.txt
      writing requirements to src/EnBiD_ananke.egg-info/requires.txt
      writing top-level names to src/EnBiD_ananke.egg-info/top_level.txt
      reading manifest file 'src/EnBiD_ananke.egg-info/SOURCES.txt'
      adding license file 'LICENSE'
      writing manifest file 'src/EnBiD_ananke.egg-info/SOURCES.txt'
      /tmp/pip-build-env-c6wiibpi/overlay/lib/python3.9/site-packages/setuptools/command/build_py.py:218: _Warning: Package 'EnBiD_ananke.__license__' is absent from the `packages` configuration.
      !!

              ********************************************************************************
              ############################
              # Package would be ignored #
              ############################
              Python recognizes 'EnBiD_ananke.__license__' as an importable package[^1],
              but it is absent from setuptools' `packages` configuration.

              This leads to an ambiguous overall configuration. If you want to distribute this
              package, please make sure that 'EnBiD_ananke.__license__' is explicitly added
              to the `packages` configuration field.

              Alternatively, you can also rely on setuptools' discovery methods
              (for example by using `find_namespace_packages(...)`/`find_namespace:`
              instead of `find_packages(...)`/`find:`).

              You can read more about "package discovery" on setuptools documentation page:

              - https://setuptools.pypa.io/en/latest/userguide/package_discovery.html

              If you don't want 'EnBiD_ananke.__license__' to be distributed and are
              already explicitly excluding 'EnBiD_ananke.__license__' via
              `find_namespace_packages(...)/find_namespace` or `find_packages(...)/find`,
              you can try to use `exclude_package_data`, or `include-package-data=False` in
              combination with a more fine grained `package-data` configuration.

              You can read more about "package data files" on setuptools documentation page:

              - https://setuptools.pypa.io/en/latest/userguide/datafiles.html

              [^1]: For Python, any directory (with suitable naming) can be imported,
                    even if it does not contain any `.py` files.
                    On the other hand, currently there is no concept of package data
                    directory, all directories are treated like packages.
              ********************************************************************************

      !!
        check.warn(importable)
      creating build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0
      copying src/EnBiD_ananke/Enbid-2.0/.DS_Store -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0
      copying src/EnBiD_ananke/Enbid-2.0/COPYING -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0
      copying src/EnBiD_ananke/Enbid-2.0/COPYRIGHT -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0
      copying src/EnBiD_ananke/Enbid-2.0/Enbid -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0
      copying src/EnBiD_ananke/Enbid-2.0/periodic_lengths.txt -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0
      creating build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/Documentation
      copying src/EnBiD_ananke/Enbid-2.0/Documentation/enbid.pdf -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/Documentation
      copying src/EnBiD_ananke/Enbid-2.0/Documentation/users-guide.pdf -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/Documentation
      copying src/EnBiD_ananke/Enbid-2.0/Documentation/users-guide.tex -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/Documentation
      copying src/EnBiD_ananke/Enbid-2.0/Documentation/users-guide.tex.flc -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/Documentation
      creating build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/Examples/hernquist1_small
      copying src/EnBiD_ananke/Enbid-2.0/Examples/hernquist1_small/snapshot_ici -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/Examples/hernquist1_small
      copying src/EnBiD_ananke/Enbid-2.0/Examples/hernquist1_small/snapshot_ici_typelist -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/Examples/hernquist1_small
      creating build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/Examples/uniform_6d_box_4
      copying src/EnBiD_ananke/Enbid-2.0/Examples/uniform_6d_box_4/snapshot_ici -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/Examples/uniform_6d_box_4
      copying src/EnBiD_ananke/Enbid-2.0/Examples/uniform_6d_box_4/snapshot_ici_3d.ascii -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/Examples/uniform_6d_box_4
      copying src/EnBiD_ananke/Enbid-2.0/Examples/uniform_6d_box_4/snapshot_ici_6d.ascii -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/Examples/uniform_6d_box_4
      creating build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/parameterfiles
      copying src/EnBiD_ananke/Enbid-2.0/parameterfiles/myparameterfile1 -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/parameterfiles
      copying src/EnBiD_ananke/Enbid-2.0/parameterfiles/myparameterfile2 -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/parameterfiles
      copying src/EnBiD_ananke/Enbid-2.0/parameterfiles/myparameterfile3 -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/parameterfiles
      copying src/EnBiD_ananke/Enbid-2.0/parameterfiles/myparameterfile4 -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/parameterfiles
      creating build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src                                                                                                                                                                         [115/1692]
      copying src/EnBiD_ananke/Enbid-2.0/src/Makefile -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src
      copying src/EnBiD_ananke/Enbid-2.0/src/allocate.cpp -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src
      copying src/EnBiD_ananke/Enbid-2.0/src/allocate.o -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src
      copying src/EnBiD_ananke/Enbid-2.0/src/allvars.cpp -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src
      copying src/EnBiD_ananke/Enbid-2.0/src/allvars.h -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src
      copying src/EnBiD_ananke/Enbid-2.0/src/allvars.o -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src
      copying src/EnBiD_ananke/Enbid-2.0/src/begrun.cpp -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src
      copying src/EnBiD_ananke/Enbid-2.0/src/begrun.o -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src
      copying src/EnBiD_ananke/Enbid-2.0/src/density_nd.cpp -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src
      copying src/EnBiD_ananke/Enbid-2.0/src/density_nd.o -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src
      copying src/EnBiD_ananke/Enbid-2.0/src/functions.cpp -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src
      copying src/EnBiD_ananke/Enbid-2.0/src/functions.h -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src
      copying src/EnBiD_ananke/Enbid-2.0/src/functions.o -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src
      copying src/EnBiD_ananke/Enbid-2.0/src/init.cpp -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src
      copying src/EnBiD_ananke/Enbid-2.0/src/init.o -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src
      copying src/EnBiD_ananke/Enbid-2.0/src/io.cpp -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src
      copying src/EnBiD_ananke/Enbid-2.0/src/io.o -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src
      copying src/EnBiD_ananke/Enbid-2.0/src/main.cpp -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src
      copying src/EnBiD_ananke/Enbid-2.0/src/main.o -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src
      copying src/EnBiD_ananke/Enbid-2.0/src/ngb_search.cpp -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src
      copying src/EnBiD_ananke/Enbid-2.0/src/ngb_search.h -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src
      copying src/EnBiD_ananke/Enbid-2.0/src/ngb_search.o -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src
      copying src/EnBiD_ananke/Enbid-2.0/src/nr.cpp -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src
      copying src/EnBiD_ananke/Enbid-2.0/src/nr.h -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src
      copying src/EnBiD_ananke/Enbid-2.0/src/nr.o -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src
      copying src/EnBiD_ananke/Enbid-2.0/src/proto.h -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src
      copying src/EnBiD_ananke/Enbid-2.0/src/read_ic.cpp -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src
      copying src/EnBiD_ananke/Enbid-2.0/src/read_ic.o -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src
      copying src/EnBiD_ananke/Enbid-2.0/src/tree.cpp -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src
      copying src/EnBiD_ananke/Enbid-2.0/src/tree.h -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src
      copying src/EnBiD_ananke/Enbid-2.0/src/tree.o -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src
      copying src/EnBiD_ananke/Enbid-2.0/src/tree_search.cpp -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src
      copying src/EnBiD_ananke/Enbid-2.0/src/tree_search.o -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/Enbid-2.0/src
      creating build/lib.linux-x86_64-cpython-39/EnBiD_ananke/__license__
      copying src/EnBiD_ananke/__license__/LICENSE -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/__license__
      copying src/EnBiD_ananke/__license__/__init__.py -> build/lib.linux-x86_64-cpython-39/EnBiD_ananke/__license__
      UPDATING build/lib.linux-x86_64-cpython-39/EnBiD_ananke/_version.py
      set build/lib.linux-x86_64-cpython-39/EnBiD_ananke/_version.py to '0.2.0b3'
      running build_ext
      building '' extension
      gcc -shared -L/rds/bear-apps/2021a/EL8-ice/software/OpenSSL/1.1/lib64 -L/rds/bear-apps/2021a/EL8-ice/software/OpenSSL/1.1/lib -L/rds/bear-apps/2021a/EL8-ice/software/libffi/3.3-GCCcore-10.3.0/lib64 -L/rds/bear-apps/2021a/EL8-ice/software/libffi/3.3
-GCCcore-10.3.0/lib -L/rds/bear-apps/2021a/EL8-ice/software/GMP/6.2.1-GCCcore-10.3.0/lib64 -L/rds/bear-apps/2021a/EL8-ice/software/GMP/6.2.1-GCCcore-10.3.0/lib -L/rds/bear-apps/2021a/EL8-ice/software/XZ/5.2.5-GCCcore-10.3.0/lib64 -L/rds/bear-apps/2021a/E
L8-ice/software/XZ/5.2.5-GCCcore-10.3.0/lib -L/rds/bear-apps/2021a/EL8-ice/software/SQLite/3.35.4-GCCcore-10.3.0/lib64 -L/rds/bear-apps/2021a/EL8-ice/software/SQLite/3.35.4-GCCcore-10.3.0/lib -L/rds/bear-apps/2021a/EL8-ice/software/ncurses/6.2-GCCcore-10
.3.0/lib64 -L/rds/bear-apps/2021a/EL8-ice/software/ncurses/6.2-GCCcore-10.3.0/lib -L/rds/bear-apps/2021a/EL8-ice/software/libreadline/8.1-GCCcore-10.3.0/lib64 -L/rds/bear-apps/2021a/EL8-ice/software/libreadline/8.1-GCCcore-10.3.0/lib -L/rds/bear-apps/202
1a/EL8-ice/software/zlib/1.2.11-GCCcore-10.3.0/lib64 -L/rds/bear-apps/2021a/EL8-ice/software/zlib/1.2.11-GCCcore-10.3.0/lib -L/rds/bear-apps/2021a/EL8-ice/software/bzip2/1.0.8-GCCcore-10.3.0/lib64 -L/rds/bear-apps/2021a/EL8-ice/software/bzip2/1.0.8-GCCco
re-10.3.0/lib -L/rds/bear-apps/2021a/EL8-ice/software/binutils/2.36.1-GCCcore-10.3.0/lib64 -L/rds/bear-apps/2021a/EL8-ice/software/binutils/2.36.1-GCCcore-10.3.0/lib -L/rds/bear-apps/2021a/EL8-ice/software/Rust/1.52.1-GCCcore-10.3.0/lib64 -L/rds/bear-app
s/2021a/EL8-ice/software/Rust/1.52.1-GCCcore-10.3.0/lib -L/rds/bear-apps/2021a/EL8-ice/software/GCCcore/10.3.0/lib64 -L/rds/bear-apps/2021a/EL8-ice/software/GCCcore/10.3.0/lib -L/rds/bear-apps/2021a/EL8-ice/software/OpenSSL/1.1/lib64 -L/rds/bear-apps/202
1a/EL8-ice/software/OpenSSL/1.1/lib -L/rds/bear-apps/2021a/EL8-ice/software/libffi/3.3-GCCcore-10.3.0/lib64 -L/rds/bear-apps/2021a/EL8-ice/software/libffi/3.3-GCCcore-10.3.0/lib -L/rds/bear-apps/2021a/EL8-ice/software/GMP/6.2.1-GCCcore-10.3.0/lib64 -L/rd
s/bear-apps/2021a/EL8-ice/software/GMP/6.2.1-GCCcore-10.3.0/lib -L/rds/bear-apps/2021a/EL8-ice/software/XZ/5.2.5-GCCcore-10.3.0/lib64 -L/rds/bear-apps/2021a/EL8-ice/software/XZ/5.2.5-GCCcore-10.3.0/lib -L/rds/bear-apps/2021a/EL8-ice/software/SQLite/3.35.
4-GCCcore-10.3.0/lib64 -L/rds/bear-apps/2021a/EL8-ice/software/SQLite/3.35.4-GCCcore-10.3.0/lib -L/rds/bear-apps/2021a/EL8-ice/software/ncurses/6.2-GCCcore-10.3.0/lib64 -L/rds/bear-apps/2021a/EL8-ice/software/ncurses/6.2-GCCcore-10.3.0/lib -L/rds/bear-ap
ps/2021a/EL8-ice/software/libreadline/8.1-GCCcore-10.3.0/lib64 -L/rds/bear-apps/2021a/EL8-ice/software/libreadline/8.1-GCCcore-10.3.0/lib -L/rds/bear-apps/2021a/EL8-ice/software/zlib/1.2.11-GCCcore-10.3.0/lib64 -L/rds/bear-apps/2021a/EL8-ice/software/zli
b/1.2.11-GCCcore-10.3.0/lib -L/rds/bear-apps/2021a/EL8-ice/software/bzip2/1.0.8-GCCcore-10.3.0/lib64 -L/rds/bear-apps/2021a/EL8-ice/software/bzip2/1.0.8-GCCcore-10.3.0/lib -L/rds/bear-apps/2021a/EL8-ice/software/binutils/2.36.1-GCCcore-10.3.0/lib64 -L/rd
s/bear-apps/2021a/EL8-ice/software/binutils/2.36.1-GCCcore-10.3.0/lib -L/rds/bear-apps/2021a/EL8-ice/software/Rust/1.52.1-GCCcore-10.3.0/lib64 -L/rds/bear-apps/2021a/EL8-ice/software/Rust/1.52.1-GCCcore-10.3.0/lib -L/rds/bear-apps/2021a/EL8-ice/software/
GCCcore/10.3.0/lib64 -L/rds/bear-apps/2021a/EL8-ice/software/GCCcore/10.3.0/lib -L/rds/bear-apps/2021a/EL8-ice/software/Python/3.9.5-GCCcore-10.3.0/lib -o build/lib.linux-x86_64-cpython-39/.cpython-39-x86_64-linux-gnu.so
      gcc: fatal error: no input files
      compilation terminated.
      error: command '/rds/bear-apps/2021a/EL8-ice/software/GCCcore/10.3.0/bin/gcc' failed with exit code 1
      [end of output]

  ERROR: Failed building wheel for EnBiD-ananke
  Building wheel for Galaxia-ananke (pyproject.toml) ... error
  error: subprocess-exited-with-error

  × Building wheel for Galaxia-ananke (pyproject.toml) did not run successfully.
  │ exit code: 1
  ╰─> [84 lines of output]
      Warning: 'classifiers' should be a list, got type 'tuple'
      running bdist_wheel
      running build
      running build_py
      creating build/lib.linux-x86_64-cpython-39/Galaxia_ananke
      copying src/Galaxia_ananke/Input.py -> build/lib.linux-x86_64-cpython-39/Galaxia_ananke
      copying src/Galaxia_ananke/Output.py -> build/lib.linux-x86_64-cpython-39/Galaxia_ananke
      copying src/Galaxia_ananke/Survey.py -> build/lib.linux-x86_64-cpython-39/Galaxia_ananke
      copying src/Galaxia_ananke/__init__.py -> build/lib.linux-x86_64-cpython-39/Galaxia_ananke
      copying src/Galaxia_ananke/__metadata__.py -> build/lib.linux-x86_64-cpython-39/Galaxia_ananke
      copying src/Galaxia_ananke/_builtin_utils.py -> build/lib.linux-x86_64-cpython-39/Galaxia_ananke
      copying src/Galaxia_ananke/_constants.py -> build/lib.linux-x86_64-cpython-39/Galaxia_ananke
      copying src/Galaxia_ananke/_defaults.py -> build/lib.linux-x86_64-cpython-39/Galaxia_ananke
      copying src/Galaxia_ananke/_templates.py -> build/lib.linux-x86_64-cpython-39/Galaxia_ananke
      copying src/Galaxia_ananke/_version.py -> build/lib.linux-x86_64-cpython-39/Galaxia_ananke
      copying src/Galaxia_ananke/utils.py -> build/lib.linux-x86_64-cpython-39/Galaxia_ananke
      creating build/lib.linux-x86_64-cpython-39/Galaxia_ananke/photometry
      copying src/Galaxia_ananke/photometry/InterfaceSvoFpsDriver.py -> build/lib.linux-x86_64-cpython-39/Galaxia_ananke/photometry
      copying src/Galaxia_ananke/photometry/Isochrone.py -> build/lib.linux-x86_64-cpython-39/Galaxia_ananke/photometry
      copying src/Galaxia_ananke/photometry/IsochroneFile.py -> build/lib.linux-x86_64-cpython-39/Galaxia_ananke/photometry
      copying src/Galaxia_ananke/photometry/PhotoSystem.py -> build/lib.linux-x86_64-cpython-39/Galaxia_ananke/photometry
      copying src/Galaxia_ananke/photometry/Photometry.py -> build/lib.linux-x86_64-cpython-39/Galaxia_ananke/photometry
      copying src/Galaxia_ananke/photometry/SvoFpsDriver.py -> build/lib.linux-x86_64-cpython-39/Galaxia_ananke/photometry
      copying src/Galaxia_ananke/photometry/__init__.py -> build/lib.linux-x86_64-cpython-39/Galaxia_ananke/photometry
      running egg_info
      writing src/Galaxia_ananke.egg-info/PKG-INFO
      writing dependency_links to src/Galaxia_ananke.egg-info/dependency_links.txt
      writing requirements to src/Galaxia_ananke.egg-info/requires.txt
      writing top-level names to src/Galaxia_ananke.egg-info/top_level.txt
[…]
