### 2024-10-01 #6234 @lheckmann
@athob thank you for addressing my comments! I tried to install the new version, which works, but then it does not let me import ananke. I get the following error:

I also tried to install it outside my conda environment, which gives the same error.

```
AssertionError: Galaxia's executable /home/lea/anaconda2/envs/pyanake/.cache/Galaxia_ananke/bin/galaxia doesn't exist, which means that Galaxia_ananke's installation didn't build the backend galaxia-ananke submodule appropriately.
Please consult https://github.com/athob/py-Galaxia-ananke/#troubleshooting-installation for troubleshooting.
```
