### 2024-10-01 #6234 @athob
Thanks @lheckmann for your comment. In your directory `/home/lea/anaconda2/envs/pyanake/.cache/Galaxia_ananke`, there should be a log directory with a couple of logs. Could you look in those to see if any of them show some errors? Also please let me know what version of pip and gcc you are using.

If possible, please consider opening an issue about it in the [py-Galaxia-ananke dedicated page](https://github.com/athob/py-Galaxia-ananke/issues) mentioning this review in it.
