### 2024-10-01 athob/py-ananke#11 @athob
I'm not finding a way to reproduce that error. I first thought it was due to `gcc` being version 10, but it installs alright with `gcc` v8 in my RHEL docker image (note it's actually RHEL 8.10, I haven't found a way to get a 8.8 image running - I'm also struggling to get module installed in it, so I'm relying on the minimalist python installation).

I guess you installed using the recommended `pip install git+https://github.com/athob/py-ananke@main` command, would you be okay trying to clone the repository manually first and then run `pip install .` from the `main` branch? I would like to know notably if `pip` does update the submodules appropriately (I believe it does because it does show that it is copying the corresponding source files), but also if the failure happens in that situation too.
