### 2024-10-01 athob/py-ananke#11 @warrickball
I originally install using `pip install git+https://github.com/athob/py-ananke@main`. I just tried to clone the repo and install with `pip install .` but get the same error. I'll have a go with my local machine (running Fedora 40) and see how it goes.
