### 2024-10-01 athob/py-ananke#11 @warrickball
For some reason it tried to build SciPy from source, so once I installed SciPy's [system-level dependencies](https://scipy.github.io/devdocs/building/index.html#system-level-dependencies), I progressed to the message I now remember seeing, which is that `enbid-ananke` requires Python<3.12.
