### 2024-10-02 #6234 @lheckmann
@athob thank you for the tip, I could solve it with the logs by installing automake.
Still made an issue in case it is something you want to add to the dependencies.
https://github.com/athob/py-Galaxia-ananke/issues/4
