### 2024-10-03 athob/py-ananke#11 @athob
The Python requirement under 3.12 is intentional at the moment, but I hope to revise it soon. To clarify, [that last comment](https://github.com/athob/py-ananke/issues/11#issuecomment-2386811231) was from your installation attempt in your local machine running Fedora 40?
