### 2024-10-03 #6234 @warrickball
Thanks @lheckmann and @rrjbca for your reviews! @athob, your submission is in essence accepted. I'll have another go at debugging my own [build issue](https://github.com/athob/py-ananke/issues/11) but also start the other editorial processes at this point. I might still request some tweaks to the code or paper.

I'm a bit pressed for time until at least 10 October but hopefully we can wrap this up by the end of next week.
