### 2024-10-03 athob/py-ananke#11 @warrickball
Sorry, yes, the local install failed purely because Fedora 40 has Python 3.12. I also hope to circle back to debugging the Python 3.9 build on my University cluster. (I'll comment on the JOSS review in that thread.)
