### 2024-10-05 athob/py-ananke#11 @warrickball
I spent some time dabbling with `py-EnBiD-ananke` in isolation, finding that `python3 setup.py build_ext` was enough to reproduce the error. I'm not used to building C/C++ extensions in Python packages but by grafting some print statements into the `ext_modules=...` part of `setup(...)`, it looks like the failing step is running at a time when EnBiD itself hasn't been downloaded, which might be why `gcc` doesn't find any files. I'm not sure if this is because each step of the installation (there seem to be at least 3) is running in an independent temporary folder.

What I did find is that if I simply commented out the `ext_modules=...` line of `setup.py`, I could install `py-EnBiD-ananke` and reproduce everything in the [testing notebook](https://py-enbid-ananke.readthedocs.io/en/main/testing_EnBiD/).

I'll try to return to this tomorrow to see if I can safely comment out `ext_modules=...` in `py-Galaxia-ananke` too, and then build `py-ananke` without issue.
