### 2024-10-05 athob/py-ananke#11 @athob
Thanks for looking into it. I'll try to test that out on my end too.

The `ext_modules` usage had been in my setup implementation ever since I started developing the wrapper. I had taken it directly from adapting the [setup installation of a package](https://github.com/GalacticDynamics-Oxford/Agama/blob/master/setup.py) I was told was also wrapping C++ source code. Back then I had played around the code, trying to remove the unnecessary and keeping the potentially necessary, to get the wrapping functionality started. But now, I can't remember why I kept that usage.
