### 2024-10-06 athob/py-ananke#11 @warrickball
Though I'll leave this open, it's peculiar enough to my particular machines that I'm going to press on with the JOSS post-review process but I'll try to chip away at this and ask some colleagues and collaborators if they've ever run into anything similar. I noticed I had `setuptools` v56.0.0 but upgrading to v75.1.0 didn't help.
