### 2024-10-06 #6234 @warrickball
@athob, I apologise for flagging this so late in the process but the paper is much too long. As mentioned in the docs on the [JOSS paper format](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain),

> The paper should be between 250-1000 words. Authors submitting papers significantly longer than 1000 words may be asked to reduce the length of their paper.

I believe the whole of the *Dependencies* section can be removed and, while I appreciate the spirit behind the *Author Contributions*, it's not typical in JOSS (and certainly not required). Note that the *Acknowledgements* are also counted so you may with to reduce those. The JOSS paper is also not intended to contain technical details beyond those relevant for a summary and statement of need. So it's good to point out that `py-ananke` is useful as a Python interface to `ananke` but probably not relevant that EnBiD and Galaxia are incorporated as git submodules.

There are also at least two places (lines 35–39 and 50–57) where the list of references is unreasonably long. You should pare those don't to a few appropriate highlights.

On more technical note, it looks like many of your references use the NASA ADS abbreviation codes (e.g. `\apj` for ApJ), which JOSS doesn't recognise, so the references are listed without journal names. Please replace these with the full journal names.

If you want to check `@editorialbot`'s word count for the paper, you can comment here with `@editorialbot check repository` (as I've done [just above](https://github.com/openjournals/joss-reviews/issues/6234#issuecomment-2395548795)).
