### 2024-10-11 59619b4
First pass through JOSS draft to cut extra - removing dependencies, author contributions & trimming framework section
