### 2024-10-11 6d98901
Second pass through JOSS draft to cut extra - more overall trimming & fixing bibliography abbreviations
