+++ joss/paper.bib
@@ -1,7 +1,7 @@
 @ARTICLE{Sanderson:2020,
        author = {{Sanderson}, Robyn E. and {Wetzel}, Andrew and {Loebman}, Sarah and {Sharma}, Sanjib and {Hopkins}, Philip F. and {Garrison-Kimmel}, Shea and {Faucher-Gigu{\`e}re}, Claude-Andr{\'e} and {Kere{\v{s}}}, Du{\v{s}}an and {Quataert}, Eliot},
         title = "{Synthetic Gaia Surveys from the FIRE Cosmological Simulations of Milky Way-mass Galaxies}",
-      journal = {\apjs},
+      journal = {The Astrophysical Journal Supplement Series},
      keywords = {Astrometry, Astronomical simulations, 80, 1857, Astrophysics - Astrophysics of Galaxies, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2020,
         month = jan,
@@ -27,7 +27,7 @@ archivePrefix = {arXiv},
 @ARTICLE{JWST:2006,
        author = {{Gardner}, Jonathan P. and {Mather}, John C. and {Clampin}, Mark and {Doyon}, Rene and {Greenhouse}, Matthew A. and {Hammel}, Heidi B. and {Hutchings}, John B. and {Jakobsen}, Peter and {Lilly}, Simon J. and {Long}, Knox S. and {Lunine}, Jonathan I. and {McCaughrean}, Mark J. and {Mountain}, Matt and {Nella}, John and {Rieke}, George H. and {Rieke}, Marcia J. and {Rix}, Hans-Walter and {Smith}, Eric P. and {Sonneborn}, George and {Stiavelli}, Massimo and {Stockman}, H.~S. and {Windhorst}, Rogier A. and {Wright}, Gillian S.},
         title = "{The James Webb Space Telescope}",
-      journal = {\ssr},
+      journal = {Space Science Reviews},
      keywords = {galaxies: formation, infrared: general, planetary systems, space vehicles: instruments, stars: formation, Astrophysics},
          year = 2006,
         month = apr,
@@ -45,7 +45,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Ferreira:2022,
        author = {{Ferreira}, Leonardo and {Adams}, Nathan and {Conselice}, Christopher J. and {Sazonova}, Elizaveta and {Austin}, Duncan and {Caruana}, Joseph and {Ferrari}, Fabricio and {Verma}, Aprajita and {Trussler}, James and {Broadhurst}, Tom and {Diego}, Jose and {Frye}, Brenda L. and {Pascale}, Massimo and {Wilkins}, Stephen M. and {Windhorst}, Rogier A. and {Zitrin}, Adi},
         title = "{Panic! at the Disks: First Rest-frame Optical Observations of Galaxy Structure at z > 3 with JWST in the SMACS 0723 Field}",
-      journal = {\apjl},
+      journal = {The Astrophysical Journal Letters},
      keywords = {Galaxy classification systems, 582, Astrophysics - Astrophysics of Galaxies},
          year = 2022,
         month = oct,
@@ -64,7 +64,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Adams:2023,
[…]
