+++ joss/paper.md
@@ -65,7 +65,7 @@ We introduce ``py-ananke``, a ``Python`` pipeline designed to generate synthetic
 
 The upcoming decade holds promise for groundbreaking discoveries, thanks to a multitude of recent and forthcoming observational facilities. The James Webb Space Telescope [@JWST:2006], for instance, with its exceptional specifications, has already delved into early universe galaxies with unprecedented detail, revealing their rich diversity [@Ferreira2022;@Adams:2023;@Finkelstein:2023;@Harikane:2023;@Casey:2023;@Eisenstein:2023]. The recently launched Euclid Telescope [@Euclid:2011] promises to shed light on the universe's accelerating expansion by surveying an immense number of galaxies [@EuclidCollaboration:2022]. The Vera Rubin Observatory [@Rubin:2019], with first light expected soon, will precisely map the Milky Way (MW) up to the virial radius and nearby galaxies, providing exceptional stellar astrometry data. Furthermore, the Nancy Grace Roman Space Telescope [@Roman:2019], set to launch in the next couple of years, will offer a wide field of view for deep-sky near-infrared exploration, facilitating the study of resolved stellar populations in nearby galaxies [@Dey:2023;@Han:2023] and our own [@Sanderson:2024]. However, these observatories will generate an unprecedented amount of raw data, necessitating community preparedness.
 
-In parallel, a number of projects have emerged over the last decade in computational astrophysics, continuously surpassing hardware and software limits to simulate galaxy formation in a cosmological context realistically [@Stinson:2010;@Stinson:2013;@Hirschmann:2014;@Dubois:2014;@Vogelsberger:2014;@Hopkins:2014;@Schaye:2015;@Crain:2015;@Khandai:2015;@Wang:2015;@Dolag:2015;@Dolag:2016;@Feng:2016;@Dave:2016;@Tremmel:2017;@Nelson:2018;@Pillepich:2018;@Springel:2018;@Henden:2018;@Hopkins:2018;@Pfeffer:2018;@Kruijssen:2019;@Peeples:2019;@Dave:2019;@Bastian:2020;@Agertz:2021;@Dubois:2021;@Applebaum:2021;@Hopkins:2023;@Rey:2023;@Feldmann:2023;@Schaye:2023;@Pakmor:2023]. These simulations serve as invaluable test beds in anticipation of the next-generation telescope era, but also for our own models. However, translating these simulations into mock observables is challenging due to the representation of stellar populations as star particles, with each particle representing a total stellar mass between $10^3$ and $10^8$ times the mass of the Sun. To compare simulations with real data, one must break down these particles into individual stars consistently. Since the simulation resolution is not "one star particle per star" in the vast majority of these simulations, producing mock observables necessarily requires a series of assumptions that can have different effects on the final prediction.
+In parallel, a number of projects have emerged over the last decade in computational astrophysics, continuously surpassing hardware and software limits to simulate galaxy formation in a cosmological context realistically [see @Crain:2023 for a recent review]. These simulations serve as invaluable test beds in anticipation of the next-generation telescope era, but also for our own models. However, translating these simulations into mock observables is challenging due to the representation of stellar populations as star particles, with each particle representing a total stellar mass between $10^3$ and $10^8$ times the mass of the Sun. To compare simulations with real data, one must break down these particles into individual stars consistently. Since the simulation resolution is not "one star particle per star" in the vast majority of these simulations, producing mock observables necessarily requires a series of assumptions that can have different effects on the final prediction.
 
[…]
