+++ joss/paper.bib
@@ -394,6 +394,23 @@ archivePrefix = {arXiv},
 ###############################
 ###############################
 
+@ARTICLE{Crain:2023,
+       author = {{Crain}, Robert A. and {van de Voort}, Freeke},
+        title = "{Hydrodynamical Simulations of the Galaxy Population: Enduring Successes and Outstanding Challenges}",
+      journal = {Annual Review of Astronomy and Astrophysics},
+     keywords = {Astrophysics - Astrophysics of Galaxies, Astrophysics - Cosmology and Nongalactic Astrophysics},
+         year = 2023,
+        month = aug,
+       volume = {61},
+        pages = {473-515},
+          doi = {10.1146/annurev-astro-041923-043618},
+archivePrefix = {arXiv},
+       eprint = {2309.17075},
+ primaryClass = {astro-ph.GA},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2023ARA&A..61..473C},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
 @ARTICLE{Stinson:2010,
        author = {{Stinson}, G.~S. and {Bailin}, J. and {Couchman}, H. and {Wadsley}, J. and {Shen}, S. and {Nickerson}, S. and {Brook}, C. and {Quinn}, T.},
         title = "{Cosmological galaxy formation simulations using smoothed particle hydrodynamics}",
