### 2024-10-11 #6234 @athob
Hi @warrickball, thanks for starting the editorial process. I looked into trimming down the manuscript accordingly with your comments.

I notably removed entirely the dependencies and author contribution section. I trimmed significantly the framework section, though kept a brief mention to the submodules due to those being part of the submission as a whole. I reviewed the general content to see where I could shorten various phrasing I used, or removed unnecessary text. I also trimmed the 2 places with a large number of references, in one case limiting to only highlights (the JWST references) and in the other, replacing entirely by a citation to a recent review of the field (hydro-simulations references). Lastly, I replaced all ADS made abbreviations by their corresponding journal names.

Regarding the acknowledgements, we can't trim it down more than it already is, a lot of is text that we are requested to quote. I independently checked for the word count of that section and it roughly account for 300~ words. Would it be okay to keep it as it is in its current version, and ignore its word count?

I also tried to make sense of what exactly the wordcount consider content to count, and noticed that it also counts all the metadata section which surprised me a bit. This metadata section roughly account for 250~ words, if I ignore the metadata and acknowledgements, the rest of the text would be slightly above the 1000 words limits. Given the latter, would it be acceptable to keep the trimmed manuscript in its latest format?
