### 2024-10-11 #6234 @warrickball
I'm happy with the word count of 1562 that you've gotten the paper down to and thanks very much for your work towards doing so. I appreciate that I should've flagged this much earlier in the editorial process!

I have a few other things to look at but should be able to revisit this submission in a few hours time. I'll generate the post-review checklist in a moment.
