### 2024-10-11 51ad951
Fix Ferreira et al. 2022 reference and remove duplicate ASCL references
