### 2024-10-12 94a0044
Fix typo in acknowledgements
