### 2024-10-12 #6234 @warrickball
I caught a final few editorial fixes in [PR #13](https://github.com/athob/py-ananke/pull/13), after which I think the paper is ready.

The next step is then for you to create a tagged release and archive and upload it to an archiving service like [Zenodo](https://zenodo.org/) or [Figshare](https://figshare.com/), and post the version number and archive DOI here so I can attach them to our metadata.

Note the requirements in [the post-review checklist above](https://github.com/openjournals/joss-reviews/issues/6234#issuecomment-2407884752), specifically that the archive authors and titles must be the same as the JOSS article. (The metadata can be fixed after the archive is created but the archive contents can't be changed without creating a new one.)

You're welcome to do it however you like but when I published in JOSS, I followed [instructions like these](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content) to publish from GitHub to Zenodo. I've never tried Figshare but [something similar looks possible](https://help.figshare.com/article/how-to-connect-figshare-with-your-github-account).
