### 2024-10-12 #6234 @athob
Thanks @warrickball or your last editorial fixes. I merged the last one just now. I have contacted the co-authors to double check that all is correct. On Tuesday at latest I'll finalize everything.

I will most likely put the archive on Zenodo. I wanted to ask if it's okay that I keep a version number in the 0.x.x at this time? I would like to reserve the 1.0.0 for when I'll finally have the software available on the PyPI after my PEP 541 request is approved.
