### 2024-10-12 #6234 @warrickball
It only has to be a new release, not necessarily v1.0.0+, but it mustn't be `alpha`, `beta`, `b` etc. I've just double checked some recent publications and there are versions like v0.4 and 0.5 around. I can replace the version in the metadata with whatever you use for the Zenodo archive, which I presume will be v0.2.0 or maybe v0.3.0.
