### 2024-10-12 #6234 @warrickball
@lheckmann Apologies for drawing you back in but could you either check off all the items in [your review checklist](https://github.com/openjournals/joss-reviews/issues/6234#issuecomment-1926571614) or confirm that it's okay for me to do so on your behalf?
