### 2024-10-17 #6234 @athob
Hi @warrickball, I am working on the zenodo archive, and created a release with its DOI at https://doi.org/10.5281/zenodo.13942955. I double checked the title, and the licence is also correct, I just need to update the author list. I had a question on that topic: zenodo distinguish creators from contributors, does it matter regarding the matching of the JOSS author list? I.e. do I put everyone as creators, or is it acceptable to put all co-authors as contributors?
