### 2024-10-21 #6234 @warrickball
> I.e. do I put everyone as creators, or is it acceptable to put all co-authors as contributors?

I don't think it matters but I've asked the editorial team and will get back to you in a day or two. The Zenodo metadata can be changed without having to create a new submission, so we can press on anyway.
