### 2024-10-22 #6234 @warrickball
I've now put my AEiC hat on and am proceeding towards publication. I'll start ticking off a few things purely editorial things.
