### 2024-10-22 #6234 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
