### 2024-10-22 4ca72c7
Editorial tweaks for JOSS publication
