+++ joss/paper.bib
@@ -1,6 +1,6 @@
 @ARTICLE{Sanderson:2020,
        author = {{Sanderson}, Robyn E. and {Wetzel}, Andrew and {Loebman}, Sarah and {Sharma}, Sanjib and {Hopkins}, Philip F. and {Garrison-Kimmel}, Shea and {Faucher-Gigu{\`e}re}, Claude-Andr{\'e} and {Kere{\v{s}}}, Du{\v{s}}an and {Quataert}, Eliot},
-        title = "{Synthetic Gaia Surveys from the FIRE Cosmological Simulations of Milky Way-mass Galaxies}",
+        title = {Synthetic {Gaia} Surveys from the {FIRE} Cosmological Simulations of {Milky Way}-mass Galaxies},
       journal = {The Astrophysical Journal Supplement Series},
      keywords = {Astrometry, Astronomical simulations, 80, 1857, Astrophysics - Astrophysics of Galaxies, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2020,
@@ -44,7 +44,7 @@ archivePrefix = {arXiv},
 
 @ARTICLE{Ferreira:2022,
        author = {{Ferreira}, Leonardo and {Adams}, Nathan and {Conselice}, Christopher J. and {Sazonova}, Elizaveta and {Austin}, Duncan and {Caruana}, Joseph and {Ferrari}, Fabricio and {Verma}, Aprajita and {Trussler}, James and {Broadhurst}, Tom and {Diego}, Jose and {Frye}, Brenda L. and {Pascale}, Massimo and {Wilkins}, Stephen M. and {Windhorst}, Rogier A. and {Zitrin}, Adi},
-        title = "{Panic! at the Disks: First Rest-frame Optical Observations of Galaxy Structure at z > 3 with JWST in the SMACS 0723 Field}",
+        title = {Panic! at the Disks: First Rest-frame Optical Observations of Galaxy Structure at z > 3 with {JWST} in the {SMACS 0723 Field}},
       journal = {The Astrophysical Journal Letters},
      keywords = {Galaxy classification systems, 582, Astrophysics - Astrophysics of Galaxies},
          year = 2022,
@@ -81,7 +81,7 @@ archivePrefix = {arXiv},
 
 @ARTICLE{Finkelstein:2023,
[…]
