### 2024-10-22 #6234 @warrickball
I've opened [a final PR](https://github.com/athob/py-ananke/pull/14) with some last tweaks. With that merged we should be ready to go after another dry run!
