### 2024-10-23 #6234 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
