### 2024-10-25 #6234 @athob
@warrickball I would like to update the preprint on the corresponding arXiv publication, it it okay to do so? ArXiv also doesn't let me simply load the final PDF. I saw that the editorialbot has a generate preprint command: would this provide me with the corresponding tex source file that produces the final PDF?
