### 2024-10-31 #6234 @warrickball
I apologise for letting this slip off my radar @athob. I was also aware of the `@editorial generate preprint` command but wasn't sure what it actually generated. Hopefully that LaTeX is what you need!
